% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{adhocuncac}
\alias{adhocuncac}
\title{net 8}
\format{\preformatted{'data.frame':	9128 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  2 2 2 3 3 3 3 3 3 3 ...
 $ alter : num  78 140 181 5 7 9 10 11 16 17 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
adhocuncac
}
\description{
net 8
}
\keyword{datasets}

