% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adjByTime}
\alias{adjByTime}
\title{create an array of adjacency matrices from dynamic data}
\usage{
adjByTime(data, n = n, Time = 10, type = "dynamic")
}
\arguments{
\item{data}{4-column data frame with year, ego, alter, and tie value (or 0 & 1)}

\item{n}{number of nodes in network. Default is n.}

\item{Time}{Time is the number of timepoints you have data for}

\item{type}{kind of network data---static, dynamic, or cumulative (tie at current timepoint if it was present at any previous timepoint)}
}
\value{
Returns an #nodes x #nodes x #timepoints array of adjacency matrices
}
\description{
create an array of adjacency matrices from dynamic data
}

