% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adjByTimee}
\alias{adjByTimee}
\title{create an array of adjacency matrices from dynamic data}
\usage{
adjByTimee(data, n = max(data[, 2:3]), Time = max(data[, 1]),
  type = "dynamic", nodeids = unique(data[, 2]))
}
\arguments{
\item{data}{a 2 - 4 column data frame or matrix with columns ordered with for timepoint, id, alter, and tie
value (optional). If data is static, a 2-3 column dataset is expected with columns ordered id, alter, and
tie value (optional).}

\item{n}{The highest node id number in the network (numeric). Default is the maximum id present in the id or
alter column (columns 2 & 3).}

\item{Time}{Time is the number of timepoints you have data for (numeric, whole numbers). Default is the
maximum of column 1 of your data}

\item{type}{kind of network data (string) ---static, dynamic, or cumulative (tie at current timepoint if it
was present at any previous timepoint)}

\item{nodeids}{Numeric vector of id numbers to be retained}
}
\value{
Returns an #nodes x #nodes x #timepoints array of adjacency matrices
}
\description{
create an array of adjacency matrices from dynamic data
}

