% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{adjmat_to_edgelist}
\alias{adjmat_to_edgelist}
\alias{adjmat_to_edgelist.matrix}
\title{From adjacency matrix to edgelist}
\usage{
adjmat_to_edgelist(...)

\method{adjmat_to_edgelist}{matrix}(adjmat, undirected = TRUE)
}
\arguments{
\item{adjmat}{Square matrix. An adjacency matrix.}

\item{undirected}{Logical. TRUE when the graph is undirected.}
}
\value{
A nx2 matrix representing an edgelist.
}
\description{
From adjacency matrix to edgelist
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Method for matrix
}}
\seealso{
\code{\link{edgelist_to_adjmat}}
}

