% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adjmatbuild}
\alias{adjmatbuild}
\title{turns a linklist into an adjacency matrix}
\usage{
adjmatbuild(dyad, n = n, nodeids)
}
\arguments{
\item{dyad}{a linklist with 3 columns (ego, alter, tie value (or 1 and 0 if tie is only present or not))}

\item{n}{n is the number of nodes in the network. Default value is n.}
}
\value{
an adjacency matrix (#nodes by #nodes, value for ties or 1 & 0 throughout)
}
\description{
turns a linklist into an adjacency matrix
}

