% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adjmatbuilde}
\alias{adjmatbuilde}
\title{turns a linklist into an adjacency matrix}
\usage{
adjmatbuilde(dyad, n = max(dyad[, 1:2]), nodeids = unique(dyad[, 1]))
}
\arguments{
\item{dyad}{a linklist with 2-3 columns (ego, alter, and optional tie value)}

\item{n}{}

\item{nodeids}{the id #s retained in the matrix. These will be indicated by the row and column names}
}
\value{
an adjacency matrix (#nodes by #nodes, value for ties or 1 & 0 throughout)
}
\description{
turns a linklist into an adjacency matrix
}

