% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{adopt_mat}
\alias{adopt_mat}
\alias{adopt_mat.integer}
\alias{adopt_mat.numeric}
\title{Creates an adoption matrices}
\usage{
adopt_mat(time, ...)

\method{adopt_mat}{integer}(time, ...)

\method{adopt_mat}{numeric}(time, ...)
}
\arguments{
\item{time}{Integer vector containing time of adoption of the innovation.}

\item{...}{Ignored.}
}
\value{
A list of 2 n x T matrices with times of adoption.
\code{cumadopt} has 1's for all years in which a node indicates having the innovation.
\code{adopt} has 1's only for the year of adoption and 0 for the rest.
}
\description{
Creates an adoption matrices
}

