% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adoptMat}
\alias{adoptMat}
\title{creates matrices indicating the adoption of innovation by year}
\usage{
adoptMat(yearObject, id = FALSE)
}
\arguments{
\item{yearObject}{vector of year of adoption in order of node ID. Must have all nodes in it, whether adopted or not.}
}
\value{
A list of matrices of dim #nodes x # years indicating for each node what year they adopted
an innovation by. Adopt_mat has 1's for all years in which a node indicates having the innovation.
Adopt_mat1 has a zero only for the year of adoption and zeros for all other years.
}
\description{
creates matrices indicating the adoption of innovation by year
}

