% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdiffusr_classes.R
\name{as_diffusionnet}
\alias{as_diffusionnet}
\alias{as_diffusionnet.array}
\alias{as_diffusionnet.matrix}
\title{Creates a \code{diffusionnet} class object}
\usage{
as_diffusionnet(graph, ...)

as_diffusionnet.matrix(graph, toa, recode = TRUE, ...)

as_diffusionnet.array(graph, toa, recode = TRUE, ...)
}
\arguments{
\item{graph}{Either an adjacency matrix, an array or an edgelist}
}
\description{
Creates a \code{diffusionnet} class object
}

