% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bilataid}
\alias{bilataid}
\title{net 16}
\format{\preformatted{'data.frame':	25304 obs. of  4 variables:
 $ year  : num  1 2 3 4 5 6 7 8 9 10 ...
 $ id    : num  9 9 9 9 9 9 9 9 9 9 ...
 $ alter : num  1 1 1 1 1 1 1 1 1 1 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
bilataid
}
\description{
net 16
}
\keyword{datasets}

