% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bilatinvesttreaty}
\alias{bilatinvesttreaty}
\title{net 17}
\format{\preformatted{'data.frame':	29000 obs. of  4 variables:
 $ year  : int  1 2 3 4 5 6 7 8 9 10 ...
 $ id    : int  1 1 1 1 1 1 1 1 1 1 ...
 $ alter : int  65 65 65 65 65 65 65 65 65 65 ...
 $ dichot: int  0 0 0 1 1 1 1 1 1 1 ...
}}
\usage{
bilatinvesttreaty
}
\description{
net 17
}
\keyword{datasets}

