% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_net}
\alias{boot_net}
\alias{print.diffnet_boot}
\title{Bootstrapping}
\usage{
boot_net(graph, statistic, R, rewire.args = list(p = 1, undirected =
  getOption("diffnet.undirected")), ncpus = getOption("boot.ncpus", 1L),
  cl = NULL)

\method{print}{diffnet_boot}(x, ...)
}
\arguments{
\item{graph}{A \code{\link{diffnet}} graph.}

\item{statistic}{A function that returns either a scalar or a vector.}

\item{R}{Integer scalar. Number of repetitions.}

\item{rewire.args}{List. Arguments to be passed to \code{\link{rewire_graph}}}

\item{ncpus}{Integer scalar. Number of CPU, passed to \code{\link[parallel:makeCluster]{makeCluster}}
from the \pkg{parallel} package.}

\item{cl}{An object of class \code{c("SOCKcluster", "cluster")}}

\item{x}{A \code{diffnet_boot} class object.}

\item{...}{Ignored.}
}
\value{
An object of class \code{diffnet_bot}. As in \code{boot}
function of the \pkg{boot} package, a list with the following elements:
\item{t0}{The observed value of statistic applied to data.}
\item{t}{A matrix with R rows each of which is a bootstrap replicate of the result of calling statistic.}
\item{R}{The value of R as passed to \code{boot_net}.}
\item{graph}{The graph passed to \code{boot_net}.}
\item{seed}{The value of \code{.Random.seed} wheb \code{boot_net} started to work.}
\item{statistic}{The function \code{statistic} passed to \code{boot_net}}
}
\description{
Bootstrapping
}

