% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet.r
\name{bootnet}
\alias{bootnet}
\alias{c.diffnet_bootnet}
\alias{hist.diffnet_bootnet}
\alias{print.diffnet_bootnet}
\alias{resample_graph}
\title{Network Bootstrapping}
\usage{
resample_graph(graph, self = NULL, useR = FALSE, ...)

bootnet(graph, statistic, R, resample.args = list(self = FALSE), ...)

\method{c}{diffnet_bootnet}(..., recursive = FALSE)

\method{print}{diffnet_bootnet}(x, ...)

\method{hist}{diffnet_bootnet}(x,
  main = "Empirical Distribution of Statistic", xlab = expression(Values ~
  of ~ t), breaks = 20, annotated = TRUE, b0 = expression(atop(plain("")
  \%up\% plain("")), t[0]), b = expression(atop(plain("") \%up\% plain("")),
  t[]), ...)
}
\arguments{
\item{graph}{Any class of accepted graph format (see \code{\link{netdiffuseR-graphs}}).}

\item{self}{Logical scalar. When \code{TRUE} allows loops (self edges).}

\item{useR}{Logical scalar. When \code{TRUE}, autolinks are filled using an
\R based rutine. Otherwise it uses the \pkg{Rcpp} implementation (default).
This is intended for testing only.}

\item{...}{Further arguments passed to the method (see details).}

\item{statistic}{An statistic}

\item{R}{Number of reps}

\item{resample.args}{List. Arguments to be passed to \code{\link{resample_graph}}}

\item{recursive}{Ignored}

\item{x}{A \code{diffnet_bootnet} class object.}

\item{main}{Character scalar. Title of the histogram.}

\item{xlab}{Character scalar. x-axis label.}

\item{breaks}{Passed to \code{\link{hist}}.}

\item{annotated}{Logical scalar. When TRUE marks the observed data average and the simulated data average.}

\item{b0}{Character scalar. When \code{annotated=TRUE}, label for the value of \code{b0}.}

\item{b}{Character scalar. When \code{annotated=TRUE}, label for the value of \code{b}.}
}
\value{
A list of class \code{diffnet_bootnet} containing the following:
\item{graph}{The graph passed to \code{bootnet}.}
\item{p.value}{The resulting p-value of the test (see details).}
\item{t0}{The observed value of the statistic.}
\item{mean_t}{The average value of the statistic applied to the simulated networks.}
\item{R}{Number of simulations.}
\item{statistic}{The function \code{statistic} passed to \code{bootnet}.}
\item{boot}{A \code{boot} class object as return from the call to \code{boot}.}
\item{resample.args}{The list \code{resample.args} passed to \code{bootnet}.}
}
\description{
Implements the bootstrapping method described in Snijders and Borgatti (1999).
This function is essentially a wrapper of \code{\link[boot:boot]{boot}}.
}
\examples{
#
set.seed(13)
g <- rgraph_ba(t=99)

ans <- bootnet(g, function(w, ...) length(w@x), R=100)
ans
}
\references{
Snijders, T. A. B., & Borgatti, S. P. (1999). Non-Parametric
Standard Errors and Tests for Network Statistics. Connections, 22(2), 1–10.
Retrieved from \url{https://insna.org/PDF/Connections/v22/1999_I-2_61-70.pdf}
}
\seealso{
Other Functions for inference: \code{\link{struct_test}}
}

