% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{brfarmersDiffNet}
\alias{brfarmersDiffNet}
\title{\code{diffnet} version of the Brazilian Farmers data}
\format{A \code{\link{diffnet}} class object.}
\description{
A directed dynamic graph with 692 vertices and 21 time periods. The attributes
in the graph are static and described in \code{\link{brfarmers}}.
}
\seealso{
Other diffusion datasets: \code{\link{brfarmers}},
  \code{\link{diffusion-data}},
  \code{\link{fakeDynEdgelist}},
  \code{\link{fakeEdgelist}}, \code{\link{fakesurveyDyn}},
  \code{\link{fakesurvey}}, \code{\link{kfamilyDiffNet}},
  \code{\link{kfamily}},
  \code{\link{medInnovationsDiffNet}},
  \code{\link{medInnovations}}
}

