% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_data.R
\name{classify_graph}
\alias{classify_graph}
\title{Analyze an R object to identify the class of graph (if any)}
\usage{
classify_graph(graph)
}
\arguments{
\item{graph}{Any class of accepted graph format (see \code{\link{netdiffuseR-graphs}}).}
}
\value{
A list of attributes including
\item{type}{Character scalar. Whether is a static or a dynamic graph}
\item{nper}{Integer scalar. Number of time periods}
\item{n}{Integer scalar. Number of vertices in the graph}
}
\description{
Analyze an R object to identify the class of graph (if any)
}
\details{
This function analyzes an R object and tries to classify it among the
accepted classes in \pkg{netdiffuseR}. If the object fails to fall in one of
the types of graphs the function returns with an error indicating what (and
when possible, where) the problem lies.
}

