% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{colony}
\alias{colony}
\title{net 5}
\format{\preformatted{'data.frame':	238 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  181 175 61 135 181 159 175 140 181 65 ...
 $ alter : num  1 2 3 5 6 7 8 8 9 10 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
colony
}
\description{
net 5
}
\keyword{datasets}

