% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{copfctc}
\alias{copfctc}
\title{net 7}
\format{\preformatted{'data.frame':	8754 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  24 36 137 184 24 36 66 158 168 175 ...
 $ alter : num  1 1 1 1 2 2 2 2 2 2 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
copfctc
}
\description{
net 7
}
\keyword{datasets}

