% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{coppollute}
\alias{coppollute}
\title{net 11}
\format{\preformatted{'data.frame':	9284 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  36 44 79 5 7 8 9 10 14 17 ...
 $ alter : num  2 2 2 3 3 3 3 3 3 3 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
coppollute
}
\description{
net 11
}
\keyword{datasets}

