% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{copuncac}
\alias{copuncac}
\title{net 9}
\format{\preformatted{'data.frame':	8026 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  14 36 53 61 78 83 88 114 124 136 ...
 $ alter : num  1 1 1 1 1 1 1 1 1 1 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
copuncac
}
\description{
net 9
}
\keyword{datasets}

