% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cumulative_adopt_count}
\alias{cumulative_adopt_count}
\title{Cummulative count of adopters}
\usage{
cumulative_adopt_count(cumadopt)
}
\arguments{
\item{cumadopt}{nxT matrix. Cumulative adoption matrix obtained from
\code{\link{toa_mat}}}
}
\description{
Cummulative count of adopters
}

