% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cumulativeAdopters}
\alias{cumulativeAdopters}
\title{calculates the percent of adopters by time period}
\usage{
cumulativeAdopters(Adopt_mat)
}
\arguments{
\item{Adopt_mat}{matrix #nodes x # timeperiods indicating if node adopted or not}
}
\value{
vector indicating the adoption rate and percent adopted by time period
}
\description{
calculates the percent of adopters by time period
}

