% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{degree}
\alias{degree}
\title{Calculates degree of each node}
\usage{
degree(adjmat, cmode = 2, undirected = TRUE, self = FALSE)
}
\arguments{
\item{adjmat}{Square matrix. Adjacency matrix}

\item{cmode}{Integer. 0 for indegree, 1 for outdegree and 2 for degree.}

\item{undirected}{Logical. TRUE when the graph is undirected.}

\item{self}{Logical. TRUE when self edges should not be considered.}
}
\value{
A numeric vector with the degree of each node.
}
\description{
Calculates degree of each node
}
\examples{
# Creating a directed graph
graph <- rand_graph(undirected=FALSE)
graph

# Comparing degree measurements
 data.frame(
   In=degree(graph, 0, undirected = FALSE),
   Out=degree(graph, 1, undirected = FALSE),
   Degree=degree(graph, 2, undirected = FALSE)
 )
}

