% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{dgr}
\alias{degree}
\alias{dgr}
\alias{dgr.array}
\alias{dgr.dgCMatrix}
\alias{dgr.list}
\alias{dgr.matrix}
\alias{indegree}
\alias{outrdegree}
\title{Indegree, outdegree and degree of the vertices}
\usage{
dgr(graph, cmode = "degree", undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"))

\method{dgr}{matrix}(graph, cmode = "degree",
  undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"))

\method{dgr}{dgCMatrix}(graph, cmode = "degree",
  undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"))

\method{dgr}{list}(graph, cmode = "degree",
  undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"))

\method{dgr}{array}(graph, cmode = "degree",
  undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"))
}
\arguments{
\item{graph}{Either an \eqn{n\times n}{n * n} Matrix or a \eqn{n\times n\times T}{n * n * T} array}

\item{cmode}{Character. Either "indegree", "outdegree" or "degree".}

\item{undirected}{Logical. TRUE when the graph is undirected.}

\item{self}{Logical. TRUE when self edges should not be considered.}
}
\value{
Either a numeric vector of size \eqn{n}{n} with the degree of each node (if graph is
a matrix), or a matrix of size \eqn{n\times T}{n * T}
}
\description{
Computes the requested degree measure for each node in the graph.
}
\examples{
# Creating a directed graph
graph <- rand_graph(undirected=FALSE)
graph

# Comparing degree measurements
 data.frame(
   In=dgr(graph, "indegree", undirected = FALSE),
   Out=dgr(graph, "outdegree", undirected = FALSE),
   Degree=dgr(graph, "degree", undirected = FALSE)
 )
}
\seealso{
Other statistics: \code{\link{cumulative_adopt_count}},
  \code{\link{exposure}}, \code{\link{hazard_rate}},
  \code{\link{infection}}, \code{\link{struct_equiv}},
  \code{\link{threshold}}
}
\keyword{univar}

