% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{diag_expand}
\alias{diag_expand}
\alias{diag_expand.array}
\alias{diag_expand.dgCMatrix}
\alias{diag_expand.diffnet}
\alias{diag_expand.list}
\alias{diag_expand.matrix}
\title{Creates a square matrix suitable for spatial statistics models.}
\usage{
diag_expand(...)

\method{diag_expand}{list}(g, self = getOption("diffnet.self"),
  valued = getOption("diffnet.valued"), ...)

\method{diag_expand}{diffnet}(g, self = getOption("diffnet.self"),
  valued = getOption("diffnet.valued"), ...)

\method{diag_expand}{matrix}(g, nper, self = getOption("diffnet.self"),
  valued = getOption("diffnet.valued"), ...)

\method{diag_expand}{array}(g, self = getOption("diffnet.self"),
  valued = getOption("diffnet.valued"), ...)

\method{diag_expand}{dgCMatrix}(g, nper, self = getOption("diffnet.self"),
  valued = getOption("diffnet.valued"), ...)
}
\arguments{
\item{...}{Further arguments to be passed to the method.}

\item{g}{A graph.}

\item{self}{Logical scalar. When \code{TRUE} non-zero diagonal elements are used.}

\item{valued}{Logical scalar. When \code{FALSE} \code{g} is treated as a binary graph.}

\item{nper}{Integer scalar. Number of time periods of the graph.}
}
\value{
A square matrix of class \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} of
size \code{(nnode(g)*nper)^2}
}
\description{
Creates a square matrix suitable for spatial statistics models.
}

