% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffnet-indexing.R
\name{diffnet_check_attr_class}
\alias{diffnet_check_attr_class}
\title{Infer whether \code{value} is dynamic or static.}
\usage{
diffnet_check_attr_class(value, meta)
}
\arguments{
\item{value}{Either a matrix, data frame or a list. Attribute values.}

\item{meta}{A list. A diffnet object's meta data.}
}
\value{
The value object either as a data frame (if static) or as a list
of data frames (if dynamic). If \code{value} does not follows the permitted
types of \code{\link{diffnet_index}}, then returns with error.
}
\description{
Intended for internal use only, this function is used in \code{\link{diffnet_index}}
methods.
}

