% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{distance}
\alias{distance}
\title{net 1}
\format{\preformatted{'data.frame':	36481 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  72 72 72 72 72 72 72 72 72 72 ...
 $ alter : num  72 162 173 187 145 172 7 21 24 35 ...
 $ dichot: num  1 1 1 1 0 0 1 1 1 1 ...
}}
\usage{
distance
}
\description{
net 1
}
\keyword{datasets}

