% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{drop_isolated}
\alias{drop_isolated}
\title{Drop isolated nodes from a graph}
\usage{
drop_isolated(adjmat, undirected = TRUE)
}
\arguments{
\item{adjmat}{Square matrix. An graph as an adjacency matrix.}

\item{undirected}{Logical. TRUE when the graph is undirected.}
}
\value{
A modified adjacency matrix excluding the isolated nodes.
}
\description{
Drop isolated nodes from a graph
}

