% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{edgelist_to_adjmat}
\alias{edgelist_to_adjmat}
\alias{edgelist_to_adjmat.data.frame}
\alias{edgelist_to_adjmat.matrix}
\title{Generates adjacency adjacency matrix from an edgelist}
\usage{
edgelist_to_adjmat(edgelist, ...)

\method{edgelist_to_adjmat}{data.frame}(edgelist, ...)

\method{edgelist_to_adjmat}{matrix}(edgelist, weights = NULL, times = NULL,
  simplify = TRUE, undirected = FALSE, skip.recode = FALSE,
  no.self = FALSE, no.multiple = FALSE)
}
\arguments{
\item{edgelist}{Two column matrix/data.frame in the form of ego -source- and
alter -target- (see details).}

\item{weights}{Numeric vector. Strength of ties (optional).}

\item{times}{Integer vector. Periodicity of the ties (optional).}

\item{simplify}{Logical. When TRUE and no times vector it will return an adjacency
matrix, otherwise an array of adjacency matrices.}

\item{undirected}{Logical. TRUE when the graph is undirected.}

\item{skip.recode}{Logical. FALSE when recode of nodes's ids is performed (see details).}

\item{no.self}{Logical. TRUE when self edges are excluded.}

\item{no.multiple}{Logical. TRUE when multiple edges should not be included
(see details).}

\item{adjmat}{Square matrix. An adjacency matrix.}
}
\value{
Either an adjacency matrix (if times is NULL) or an array of these
(if times is not null).
}
\description{
Generates adjacency adjacency matrix from an edgelist
}
\details{
The edgelist must be coded from 1:n (otherwise it may cause an error).
By default, the function will \code{\link{recode}} the edgelist before starting.

When multiple edges are included, each vertex between \{i,j\} will be accounted
as many times it appears in the edgelist. So if a vertex \{i,j\} appears 2
times, the adjacency matrix element (i,j) will have a 2.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for data frame

\item \code{matrix}: Method for matrix object
}}
\examples{
# Base data
set.seed(123)
n <- 10
edgelist <- matrix(sample(1:n, size = n*10, replace = TRUE), ncol=2)
times <- sample.int(10, nrow(edgelist), replace=TRUE)
w <- abs(rnorm(nrow(edgelist)))

# Simple example
edgelist_to_adjmat(edgelist)
edgelist_to_adjmat(edgelist, undirected = TRUE)

# Using weights
edgelist_to_adjmat(edgelist, w)
edgelist_to_adjmat(edgelist, w, undirected = TRUE)

# Using times
edgelist_to_adjmat(edgelist, times = times)
edgelist_to_adjmat(edgelist, times = times, undirected = TRUE)

# Using times and weights
edgelist_to_adjmat(edgelist, times = times, weights = w)
edgelist_to_adjmat(edgelist, times = times, undirected = TRUE, weights = w)
}
\seealso{
\code{\link{adjmat_to_edgelist}}
}

