% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{epigames}
\alias{epigames}
\title{Epi Games Dataset}
\format{
A list with two data frames:

**attributes**: A data frame with 594 rows and 9 variables representing nodes:
\describe{
 \item{id}{Unique identifier for the participant.}
 \item{toa}{Time of Adoption (1 to 15), representing when the individual was first infected. Non-infected individuals have `NA`.}
 \item{qyes_total}{Cumulative count of times the player participated or scored positively in informative/educational "quarantine" questionnaires.}
 \item{qno_total}{Cumulative count of times the non-quarantine questionnaire factor was registered.}
 \item{mask_prop}{Proportion of time (across 15 steps) the participant used the mask intervention (0.0 to 1.0).}
 \item{med_prop}{Proportion of time the individual used pharmacological interventions or treatments.}
 \item{group}{Experimental group or node cohort.}
 \item{final_score}{Final score obtained in the game.}
 \item{status}{Final state label ("infected" or "not_infected").}
}

**edgelist**: A longitudinal data frame with 23,684 rows and 4 variables representing edges/contacts:
\describe{
 \item{sender}{Origin node ID of the contact.}
 \item{receiver}{Destination node ID of the contact.}
 \item{time}{Time period of the contact (1 to 15).}
 \item{weight}{Strength, duration, or density of the exposure.}
}
}
\source{
WKU Epi Game simulation
}
\description{
The WKU Epi Games dataset represents a simulated epidemic or game environment with 
dynamic encounters over 15 time periods. It provides both node-level 
attributes and a longitudinal edgelist.
}
\seealso{
Other diffusion datasets: 
\code{\link{brfarmers}},
\code{\link{brfarmersDiffNet}},
\code{\link{diffusion-data}},
\code{\link{epigamesDiffNet}},
\code{\link{fakeDynEdgelist}},
\code{\link{fakeEdgelist}},
\code{\link{fakesurvey}},
\code{\link{fakesurveyDyn}},
\code{\link{kfamily}},
\code{\link{kfamilyDiffNet}},
\code{\link{medInnovations}},
\code{\link{medInnovationsDiffNet}}
}
\concept{diffusion datasets}
