% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{epigamesDiffNet}
\alias{epigamesDiffNet}
\title{\code{diffnet} version of the Epi Games data}
\format{
A \code{\link{diffnet}} class object.
}
\source{
WKU Epi Game simulation
}
\description{
A directed dynamic graph with 594 vertices and 15 time periods. The attributes
in the graph are described in \code{\link{epigames}}.
}
\details{
By default, this \code{diffnet} object is **non-cumulative** (each slice represents 
ephemeral daily contacts) and **valued** (edge weights represent contact duration in seconds). 

To reconstruct the classic cumulative/binarized network, you can run:

\preformatted{
epigames_cumul <- epigamesDiffNet

# 1. Accumulate the history across time periods
epigames_cumul$graph <- Reduce("+", epigames_cumul$graph, accumulate = TRUE)

# 2. Apply a logical cut-off to binarize the network
epigames_cumul$graph <- lapply(epigames_cumul$graph, function(m) { m@x[] <- 1; m })
}

Non-adopters have \code{toa = NA}.
}
\seealso{
Other diffusion datasets: 
\code{\link{brfarmers}},
\code{\link{brfarmersDiffNet}},
\code{\link{diffusion-data}},
\code{\link{epigames}},
\code{\link{fakeDynEdgelist}},
\code{\link{fakeEdgelist}},
\code{\link{fakesurvey}},
\code{\link{fakesurveyDyn}},
\code{\link{kfamily}},
\code{\link{kfamilyDiffNet}},
\code{\link{medInnovations}},
\code{\link{medInnovationsDiffNet}}
}
\concept{diffusion datasets}
