% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eventhistory}
\alias{eventhistory}
\title{makes event history database}
\usage{
eventhistory(exposureObject, toa, Adopt_mat, net_no, conv)
}
\arguments{
\item{exposureObject}{exposure matrices from ExposureCalc}

\item{toa}{vector indicating times of adoption}

\item{Adopt_mat}{}

\item{net_no}{number assigned to network (might remove)}

\item{conv}{convention number (might remove)}
}
\description{
makes event history database
}

