% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eventhistorye}
\alias{eventhistorye}
\title{makes event history database}
\usage{
eventhistorye(exposureObject, toa, Adopt_mat)
}
\arguments{
\item{exposureObject}{exposure matrices from ExposureCalc}

\item{toa}{vector indicating times of adoption}

\item{Adopt_mat}{}
}
\description{
makes event history database
}

