% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{exposure}
\alias{exposure}
\title{Ego exposure}
\usage{
exposure(graph, cumadopt, wtype = 0, v = 1, undirected = TRUE)
}
\arguments{
\item{graph}{An array of adjacency matrices}

\item{cumadopt}{nxT matrix. Cumulative adoption matrix obtained from
\code{\link{toa_mat}}}

\item{wtype}{Integer. Weighting type (see details).}

\item{v}{Double. Constant for Structural Equivalence.}

\item{undirected}{Logical. TRUE if the graph is undirected.}
}
\value{
A matrix of size nxT with exposure for each node.
}
\description{
Ego exposure
}

