% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExposureCalc}
\alias{ExposureCalc}
\title{exposurecalc uses network and innovation adoption data to calculate exposure to the innovation for each node}
\usage{
ExposureCalc(all_nets, Adopt_mat)
}
\arguments{
\item{all_nets}{Adjacency matrices for the dynamic network.}

\item{Adopt_mat}{Matrix indicating if a node has adopted an innovation}

\item{Time}{The timepoint that you are calculating exposure for}
}
\value{
A list of 2 exposure matrices, one through structural equivalence and the other through adjacency
}
\description{
exposurecalc uses network and innovation adoption data to calculate exposure to the innovation for each node
}

