% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExposureCalce}
\alias{ExposureCalce}
\title{exposurecalc uses network and innovation adoption data to calculate exposure to the innovation for each node}
\usage{
ExposureCalce(all_nets, Adopt_mat, method = "direct")
}
\arguments{
\item{all_nets}{Adjacency matrices for the dynamic network.}

\item{Adopt_mat}{Matrix indicating if a node has adopted an innovation}

\item{method}{Method to use for calculating exposure, options "direct", "twostep", "threestep", or "se"
(structural equivalence)}
}
\value{
A matrix with number of rows = number of nodes and number of columns = number of timepoints
indicating exposure for each node at each timepoint.
}
\description{
exposurecalc uses network and innovation adoption data to calculate exposure to the innovation for each node
}

