% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{fakeEdgelist}
\alias{fakeEdgelist}
\title{Fake static edgelist}
\format{
A data frame with 11 rows and 3 variables
\describe{
 \item{ego}{Nominating individual}
 \item{alter}{Nominated individual}
 \item{value}{Strength of the tie}
}
}
\source{
Generated for the package
}
\description{
A data frame used for examples in reading edgelist format networks. This
edgelist can be merged with the dataset \code{\link{fakesurvey}}.
}
\seealso{
Other diffusion datasets: 
\code{\link{brfarmers}},
\code{\link{brfarmersDiffNet}},
\code{\link{diffusion-data}},
\code{\link{fakeDynEdgelist}},
\code{\link{fakesurvey}},
\code{\link{fakesurveyDyn}},
\code{\link{kfamily}},
\code{\link{kfamilyDiffNet}},
\code{\link{medInnovations}},
\code{\link{medInnovationsDiffNet}}
}
\author{
George G. Vega Yon
}
\concept{diffusion datasets}
