% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fakesurvey}
\alias{fakesurvey}
\title{Fake survey data}
\format{A data frame with 9 rows and 9 variables
\describe{
 \item{id}{Unique id at group level}
 \item{toa}{Time of adoption}
 \item{group}{Group id}
 \item{net1}{Network nomination 1}
 \item{net2}{Network nomination 2}
 \item{net3}{Network nomination 3}
 \item{age}{Age of the respondent}
 \item{gender}{Gende of the respondent}
 \item{note}{Descroption of the respondent}
}}
\source{
Generated for the package.
}
\description{
This data frame is used to ilustrate some of the functions of the package,
in particular, the \code{\link{survey_to_diffnet}} function. This dataset
can be merged with the \code{\link{fakeEdgelist}}.
}
\author{
George G. Vega Yon
}
\seealso{
Other diffusion datasets: \code{\link{brfarmersDiffNet}},
  \code{\link{brfarmers}}, \code{\link{diffusion-data}},
  \code{\link{fakeDynEdgelist}},
  \code{\link{fakeEdgelist}}, \code{\link{fakesurveyDyn}},
  \code{\link{kfamilyDiffNet}}, \code{\link{kfamily}},
  \code{\link{medInnovationsDiffNet}},
  \code{\link{medInnovations}}
}

