% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{glposts}
\alias{glposts}
\title{net 13}
\format{\preformatted{'data.frame':	5843 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  48 109 77 109 65 120 83 181 94 181 ...
 $ alter : num  10 17 31 83 9 17 7 48 83 166 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
glposts
}
\description{
net 13
}
\keyword{datasets}

