% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{glreferral}
\alias{glreferral}
\title{net 12}
\format{\preformatted{'data.frame':	1330 obs. of  4 variables:
 $ year  : num  2 0 5 6 5 4 5 2 7 3 ...
 $ id    : num  165 183 31 35 183 183 35 183 35 183 ...
 $ alter : num  3 3 3 7 7 7 7 7 7 7 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
glreferral
}
\description{
net 12
}
\keyword{datasets}

