% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{glsubscription}
\alias{glsubscription}
\title{net 14}
\format{\preformatted{'data.frame':	25246 obs. of  4 variables:
 $ year  : int  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : int  31 181 183 9 17 24 31 37 44 48 ...
 $ alter : int  3 3 3 7 7 7 7 7 7 7 ...
 $ dichot: int  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
glsubscription
}
\description{
net 14
}
\keyword{datasets}

