% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hazard_rate}
\alias{hazard_rate}
\title{Calculates hazard rate}
\usage{
hazard_rate(cumadopt)
}
\arguments{
\item{cumadopt}{nxT matrix. Cumulative adoption matrix obtained from
\code{\link{toa_mat}}}
}
\value{
A vector indicating the hazard rate of each node
}
\description{
Calculates hazard rate
}

