% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hazardrate}
\alias{hazardrate}
\title{report the hazard rate}
\usage{
hazardrate(Adopt_mat)
}
\arguments{
\item{Adopt_mat}{matrix #nodes x # timeperiods indicating if node adopted or not}
}
\value{
vector of hazard rates
}
\description{
report the hazard rate
}

