% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{imfassets}
\alias{imfassets}
\title{net 18}
\format{\preformatted{'data.frame':	40418 obs. of  4 variables:
 $ year  : num  5 8 9 8 2 3 4 5 6 4 ...
 $ id    : num  1 1 1 1 1 1 1 1 1 1 ...
 $ alter : num  44 44 57 139 159 159 159 159 159 168 ...
 $ dichot: num  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
imfassets
}
\description{
net 18
}
\keyword{datasets}

