% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{inbfctc}
\alias{inbfctc}
\title{net 6}
\format{\preformatted{'data.frame':	9470 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  9 10 17 24 31 36 42 48 51 60 ...
 $ alter : num  2 2 2 2 2 2 2 2 2 2 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
inbfctc
}
\description{
net 6
}
\keyword{datasets}

