% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{incpollute}
\alias{incpollute}
\title{net 10}
\format{\preformatted{'data.frame':	7824 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  31 183 7 9 10 16 17 23 24 30 ...
 $ alter : num  2 2 3 3 3 3 3 3 3 3 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
incpollute
}
\description{
net 10
}
\keyword{datasets}

