% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect_suscept.R
\name{infection}
\alias{infection}
\alias{infection.array}
\alias{susceptibility}
\alias{susceptibility.array}
\title{Susceptibility and Infection}
\usage{
infection(graph, times, normalize = TRUE)

\method{infection}{array}(graph, times, normalize = TRUE)

susceptibility(graph, times, normalize = TRUE)

\method{susceptibility}{array}(graph, times, normalize = TRUE)
}
\arguments{
\item{graph}{Array of size \eqn{n\times n\times T}{nxnxT} as a dynamic graph}

\item{times}{Integer vector with times of adoption}

\item{normalize}{Logical. Whether or not to normalize the outcome}
}
\value{
Numeric vector of size \eqn{n}{n}.
}
\description{
Calculates infectiousness and susceptibility for each node in the graph
}
\details{
Normalization, \code{normalize=TRUE}, is applied by dividing the
resulting number by the number of individuals who had already adopted the
innovation by the time \eqn{t}{t}.

Given that node \eqn{i}{i} adopted the innovation in time \eqn{t}{t}, its
Susceptibility is calculated as follows

\deqn{S_i = \frac{%
\sum_{j=1}^n x_{ij(t)}z_{j(t-1)}}{%
\sum_{j=1}^n x_{ij(t)}z_{j(1\leq t \leq t-1)} }\qquad \mbox{for }i,j=1,\dots,n\quad i\neq j}{%
S(i) = [sum(x(ij,t) * z(j,t-1))]/[sum(x(ij,t) * z(j, 1<=t<=t-1))]}

where \eqn{x_{ij(t)}}{x(ij,t)} is 1 whenever there's a link from \eqn{i}{i}
to \eqn{j}{j} at time \eqn{t}{t}, \eqn{z_{j(t-1)}}{z(j,t-1)}
is 1 whenever individual \eqn{j}{j} adopted the innovation at time \eqn{t-1}{t-1},
and \eqn{x_{ij(t)}z_{j(1\leq t \leq t-1)}}{z(j, 1<=t<=t-1)} is 1 whenever
\eqn{j}{j} had adopted the innovation prior to \eqn{i}{i}.

In the same way, infectiousness is calculated as follows

\deqn{I_i = \frac{%
\sum_{j=1}^n x_{ji(t)}z_{j(t+1)}}{%
\sum_{j=1}^n x_{ji(t)}z_{j(t+1\leq t \leq T)} }\qquad \mbox{for }i,j=1,\dots,n\quad i\neq j}{%
I(i) = [sum(x(ji,t) * z(j,t+1))]/[sum(x(ji,t) * z(j, t+1<=t<=T))]}

It is worth noticing that, as we can see in the formulas, while susceptibility
is from alter to ego, infection is from ego to alter.

Finally, by default both are normalized by the number of individuals who had
adopted the innovation by time \eqn{t-1}{t-1}. Thus, the resulting formulas,
when \code{normalize=TRUE}, can be rewritten as

\deqn{%
S_i' = \frac{S_i}{\sum_{j=1}^nz_{j(t-1)}} \qquad I_i' = \frac{I_i}{\sum_{j=1}^nz_{j(t-1)}}}{%
S(i)' = S(i)/[sum(z(j,t-1))]

I(i)' = I(i)/[sum(z(j,t-1))]}
}
\references{
Thomas W. Valente, Stephanie R. Dyal, Kar-Hai Chu, Heather Wipfli, Kayo
Fujimoto, Diffusion of innovations theory applied to global tobacco control
treaty ratification, Social Science & Medicine, Volume 145, November 2015,
Pages 89-97, ISSN 0277-9536
(\url{http://dx.doi.org/10.1016/j.socscimed.2015.10.001})

Myers, D. J. (2000). "The Diffusion of Collective Violence: Infectiousness,
Susceptibility, and Mass Media Networks". American Journal of Sociology, 106(1),
173–208. doi:10.1086/303110
}

