% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{internatsettle}
\alias{internatsettle}
\title{net 15}
\format{\preformatted{'data.frame':	21725 obs. of  4 variables:
 $ year  : num  9 10 2 3 4 6 7 8 9 10 ...
 $ id    : num  9 9 9 9 9 9 9 9 9 9 ...
 $ alter : num  1 1 7 7 7 7 7 7 7 7 ...
 $ dichot: num  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
internatsettle
}
\description{
net 15
}
\keyword{datasets}

