% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kfamilyDiffNet}
\alias{kfamilyDiffNet}
\title{\code{diffnet} version of the Korean Family Planning data}
\format{A \code{\link{diffnet}} class object.}
\usage{
kfamilyDiffNet
}
\description{
A directed dynamic graph with 1,047 vertices and 11 time periods. The attributes
in the graph are static and described in \code{\link{kfamily}}.
}
\seealso{
Other diffusion datasets: \code{\link{brfarmersDiffNet}},
  \code{\link{brfarmers}}, \code{\link{diffusion-data}},
  \code{\link{fakeDynEdgelist}},
  \code{\link{fakeEdgelist}}, \code{\link{fakesurveyDyn}},
  \code{\link{fakesurvey}}, \code{\link{kfamily}},
  \code{\link{medInnovationsDiffNet}},
  \code{\link{medInnovations}}
}
\keyword{datasets}

