% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{kp}
\alias{kp}
\title{Korean Family Planning}
\source{
Rogers & Kincaid 1981
}
\description{
1047 respondents women
25 communities
Family Planning
Length of diffusion 11 years
Year of collection 1973
Years span
}
\references{
TOM
}

