% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{matrix_compare}
\alias{matrix_compare}
\title{Element-wise comparison between two matrices}
\usage{
matrix_compare(A, B, fun)
}
\arguments{
\item{A}{A matrix of size \code{n*m} of class \code{dgCMatrix}.}

\item{B}{A matrix of size \code{n*m} of class \code{dgCMatrix}.}

\item{fun}{A function that receives 2 arguments and returns a scalar.}
}
\value{
An object of class \code{dgCMatrix} of size \code{n*m}.
}
\description{
Element-wise comparison between two matrices
}
