% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{medInnovations}
\alias{medInnovations}
\title{Medical Innovation}
\format{A data frame with 125 rows and 59 columns:
\describe{
   \item{city}{city id}
   \item{id}{sequential respondent id}
   \item{detail}{detail man}
   \item{meet}{meetings, lectures, hospitals}
   \item{coll}{colleagues}
   \item{attend}{attend professional meets}
   \item{proage}{professional age}
   \item{length}{lenght of reside in community}
   \item{here}{only practice here}
   \item{science}{science versus patients}
   \item{position}{position in home base}
   \item{journ2}{journal subscriptions}
   \item{paadico}{Percent alter adoption date imp}
   \item{ado}{adoption month 1 to 18}
   \item{thresh}{threshold}
   \item{ctl}{corrected tl tl-exp level}
   \item{catbak}{category 1-init 2-marg 3-low tl}
   \item{sourinfo}{source of information}
   \item{origid}{original respondent id}
   \item{adopt}{adoption date 1= 11/53}
   \item{recon}{reconstructed med innov}
   \item{date}{date became aware}
   \item{info}{information source}
   \item{most}{most important info source}
   \item{journ}{journals}
   \item{drug}{drug houses}
   \item{net1_1}{advisor nomination1}
   \item{net1_2}{advisor nomination2}
   \item{net1_3}{advisor nomination3}
   \item{net2_1}{discuss nomination1}
   \item{net2_2}{discuss nomination2}
   \item{net2_3}{discuss nomination3}
   \item{net3_1}{friends nomination1}
   \item{net3_2}{friends nomination2}
   \item{net3_3}{friends nomination3}
   \item{nojourn}{number of pro journals receive}
   \item{free}{free time companions}
   \item{social}{med discussions during social}
   \item{club}{club membership}
   \item{friends}{friends are doctors}
   \item{young}{young patients}
   \item{nonpoor}{nonpoverty patients}
   \item{office}{office visits}
   \item{house}{house calls}
   \item{tend}{tendency to prescribe drugs}
   \item{reltend}{relative tendency to prescribe}
   \item{perc}{perceived drug competition}
   \item{proximty}{physical proximity to other doc}
   \item{home}{home base hospital affiliation}
   \item{special}{specialty}
   \item{belief}{belief in science}
   \item{proage2}{profesional age 2}
   \item{presc}{prescription prone}
   \item{detail2}{contact with detail man}
   \item{dichot}{dichotomous personal preference}
   \item{expect}{adoption month expected}
   \item{recall}{recalls adopting}
   \item{commun}{Number of community}
   \item{toa}{Time of Adoption}
   \item{study}{Number of study in Valente (1995)}
}}
\source{
The Medical Innovation data were stored in file cabinets in a basement
building at Columbia University. Ron Burt (1987) acquired an NSF grant to
develop network diffusion models and retrieve the original surveys and enter
them into a database. He distributed copies of the data on diskette and sent
one to me, Tom Valente, and I imported onto a PC environment.
}
\description{
From Valente (1995) \dQuote{Coleman, Katz and Menzel from Columbia University's Bureau of Applied Research
studied the adoption of tetracycline by physiciams in four Illinois communities
in 1954.[...] Tetracycline was a powerful and useful antibiotic just introduced in
the mid-1950s}
}
\details{
The collected dataset has 125 respondents (doctors), and spans 17 months of data
collected in 1955. Time of adoption of non-adopters has been set to month
18 (see the manual entry titled \code{\link[netdiffuseR:diffusion-data]{Difussion Network Datasets}}).
}
\references{
Coleman, J., Katz, E., & Menzel, H. (1966). Medical innovation: A diffusion
study (2nd ed.). New York: Bobbs-Merrill

Valente, T. W. (1995). Network models of the diffusion of innovations (2nd ed.).
Cresskill N.J.: Hampton Press.
}
\seealso{
Other diffusion datasets: \code{\link{brfarmersDiffNet}},
  \code{\link{brfarmers}}, \code{\link{diffusion-data}},
  \code{\link{fakeDynEdgelist}},
  \code{\link{fakeEdgelist}}, \code{\link{fakesurveyDyn}},
  \code{\link{fakesurvey}}, \code{\link{kfamilyDiffNet}},
  \code{\link{kfamily}},
  \code{\link{medInnovationsDiffNet}}
}
