% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{medInnovationsDiffNet}
\alias{medInnovationsDiffNet}
\title{\code{diffnet} version of the Medical Innovation data}
\format{A \code{\link{diffnet}} class object.}
\description{
A directed dynamic graph with 125 vertices and 18 time periods. The attributes
in the graph are static and described in \code{\link{medInnovations}}.
}
\seealso{
Other diffusion datasets: 
\code{\link{brfarmersDiffNet}},
\code{\link{brfarmers}},
\code{\link{diffusion-data}},
\code{\link{fakeDynEdgelist}},
\code{\link{fakeEdgelist}},
\code{\link{fakesurveyDyn}},
\code{\link{fakesurvey}},
\code{\link{kfamilyDiffNet}},
\code{\link{kfamily}},
\code{\link{medInnovations}}
}
\concept{diffusion datasets}
