% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mentor_matching}
\alias{mentor_matching}
\alias{leader_matching}
\title{Optimal Leader/Mentor Matching}
\usage{
mentor_matching(graph, n, cmode = "indegree", lead.ties.method = "average",
  geodist.args = list())

leader_matching(graph, n, cmode = "indegree", lead.ties.method = "average",
  geodist.args = list())
}
\arguments{
\item{graph}{Any class of accepted graph format (see \code{\link{netdiffuseR-graphs}}).}

\item{n}{Number of leaders}

\item{cmode}{Passed to \code{\link{dgr}}}

\item{lead.ties.method}{Passed to \code{\link{rank}}}

\item{geodist.args}{Passed to \code{\link{approx_geodesic}}.}
}
\value{
A data frame with the following columns:
\item{degree}{Numeric. Degree of each vertex in the graph}
\item{iselader}{Logical. \code{TRUE} when the vertex was picked as a leader.}
\item{match}{Integer. The corresponding matched leader.}
}
\description{
Implementes the algorithm described in Valente and Davis (1999)
}
\details{
The algorithm works
}
\examples{

set.seed(1231)
graph <- rgraph_ba(t=9, m = 2, self=FALSE)

ans <- mentor_matching(graph, n = 2)
head(ans)
ig <- igraph::graph_from_adjacency_matrix(graph)

plot(ig, vertex.color = ifelse(ans$isleader, "transparent", ans$match + 2))

}
\references{
Valente, T. W., & Davis, R. L. (1999). Accelerating the Diffusion of
Innovations Using Opinion Leaders. The ANNALS of the American Academy of
Political and Social Science, 566(1), 55–67.
\doi{doi.org/doi.org/10.1177/000271629956600105}
}
