% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{netdiffuseR-options}
\alias{netdiffuseR-options}
\title{\pkg{netdiffuseR} default options}
\value{
The full list of options follows:
\item{undirected}{FALSE}
\item{self}{FALSE}
\item{multiple}{FALSE}
\item{tol}{1e-8 (used for package testing)}
\item{valued}{FALSE}
\item{outgoing}{TRUE}
\item{keep.isolates}{TRUE}
}
\description{
\pkg{netdiffuseR} default options
}
\details{
Set of default options used by the package. These can be retrieved
via \code{\link{getOption}} using the prefix \code{diffnet} (see examples)
}
\examples{
getOption("diffnet.undirected")
getOption("diffnet.multiple")
getOption("diffnet.self")
}
\author{
George G. Vega Yon
}

