% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{netdiffuseR}
\alias{netdiffuseR}
\alias{netdiffuseR-package}
\title{netdiffuseR}
\description{
You can access to the project website at
\url{https://github.com/USCCANA/diffusiontest}
}
\details{
Network Analysis for Diffusion of Innovations
}

