% Generated by roxygen2 (4.0.2): do not edit by hand
\name{netprep}
\alias{netprep}
\title{take a stata dataset and turn it into a linklist within a 4-column dataframe}
\usage{
netprep(data, varnames = c("year", "id", "alter", "dichot"))
}
\arguments{
\item{data}{a dataset in stata format with columns for year, ego_number, alter_number, and dichot (tie or no tie)}

\item{varnames}{variable names you want the output dataframe to have, default is year, id, alter, dichot}
}
\value{
a linklist in form of dataframe with columns for year and dichot (1 or 0 indicating tie presence)
}
\description{
take a stata dataset and turn it into a linklist within a 4-column dataframe
}

