% Generated by roxygen2 (4.0.2): do not edit by hand
\name{netprepe}
\alias{netprepe}
\title{take a stata dataset and turn it into a linklist within a 4-column dataframe}
\usage{
netprepe(data, varnames = c("year", "id", "alter", "dichot"))
}
\arguments{
\item{data}{in order rawdata$year, rawdata$ego_number, rawdata$alter_number, rawdata$dicho}

\item{varnames}{variable names you want the output dataframe to have, default is year, id, alter, dichot}
}
\value{
a linklist in form of dataframe with columns for year and dichot (1 or 0 indicating tie presence)
}
\description{
take a stata dataset and turn it into a linklist within a 4-column dataframe
}

