% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.r
\name{network}
\alias{network}
\alias{diffnet_to_network}
\alias{diffnet_to_networkDynamic}
\alias{networkDynamic_to_diffnet}
\title{Convertion between graph classes}
\usage{
diffnet_to_network(graph, slices = 1:nslices(graph), ...)

diffnet_to_networkDynamic(graph, slices = 1:nslices(graph),
  diffnet2net.args = list(), netdyn.args = list())

networkDynamic_to_diffnet(graph, toavar)
}
\arguments{
\item{graph}{An object of class \code{\link{diffnet}}}

\item{slices}{An integer vector indicating the slices to subset}

\item{...}{Further arguments passed to \code{\link[networkDynamic:networkDynamic]{networkDynamic}}}

\item{diffnet2net.args}{List of arguments passed to \code{diffnet_to_network}.}

\item{netdyn.args}{List of arguments passed to \code{\link[networkDynamic:networkDynamic]{networkDynamic}}}

\item{toavar}{Character scalar. Name of the vertex attribute that holds the times of adoption.}
}
\value{
\code{diffnet_to_network} returns a list of length \code{length(slices)} in which
each element is a \code{\link[network:network]{network}} object corresponding a slice of the
\code{graph} (\code{diffnet} object). The attributes list will include \code{toa} (time of
adoption).

An object of class \code{networkDynamic}.
}
\description{
Convertion between graph classes
}
\details{
\code{diffnet_to_networkDynamic} calls \code{diffnet_to_network} and
uses the output to call \code{networkDynamic}, passing the resulting list of
\code{network} objects as \code{network.list} (see \code{\link[networkDynamic:networkDynamic]{networkDynamic}}).

By default, \code{diffnet_to_networkDynamic} passes \code{net.obs.period} as
\preformatted{
  net.obs.period = list(
    observations = list(range(graph$meta$pers)),
    mode="discrete",
    time.increment = 1,
    time.unit = "step"
  )
}

By default, \code{networkDynamic_to_diffnet} uses the first slice as reference for
vertex attributes and times of adoption.
}
\examples{
data(medInnovationsDiffNet)
diffnet_to_networkDynamic(medInnovationsDiffNet)
}
\seealso{
Other Foreign: \code{\link{igraph}},
  \code{\link{read_pajek}}, \code{\link{read_ucinet_head}}
}
