% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{phone}
\alias{phone}
\title{net 4}
\format{\preformatted{'data.frame':	37501 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  159 149 165 65 120 51 77 159 86 56 ...
 $ alter : num  139 53 53 31 67 61 15 138 161 135 ...
 $ dichot: num  1 0 1 1 1 1 0 0 1 0 ...
}}
\usage{
phone
}
\description{
net 4
}
\keyword{datasets}

