% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_adopters}
\alias{plot_adopters}
\title{Visualize adopters and cumulative adopters}
\usage{
plot_adopters(obj, freq = FALSE, what = c("adopt", "cumadopt"),
  add = FALSE, include.legend = TRUE, pch = c(21, 21), type = c("b",
  "b"), ylim = if (!freq) c(0, 1) else NULL, lty = c(1, 1),
  col = c("black", "black"), bg = c("lightblue", "gray"), xlab = "Time",
  ylab = ifelse(freq, "Frequency", "Proportion"),
  main = "Adopters and Cumulative Adopters", ...)
}
\arguments{
\item{obj}{Either a diffnet object or a cumulative a doption matrix.}

\item{freq}{Logical scalar. When TRUE frequencies are plotted instead of proportions.}

\item{what}{Character vector of length 2. What to plot.}

\item{add}{Logical scalar. When TRUE lines and dots are added to the current graph.}

\item{include.legend}{Logical scalar. When TRUE a legend of the graph is plotted.}

\item{pch}{Integer vector of length 2. See \code{\link{matplot}}.}

\item{type}{Character vector of length 2. See \code{\link{matplot}}.}

\item{ylim}{Numeric vector of length 2. Sets the plotting limit for the y-axis.}

\item{lty}{Numeric vector of length 2. See \code{\link{matplot}}.}

\item{col}{Character vector of length 2. See \code{\link{matplot}}.}

\item{bg}{Character vector of length 2. See \code{\link{matplot}}.}

\item{xlab}{Character scalar. Name of the x-axis.}

\item{ylab}{Character scalar. Name of the y-axis.}

\item{main}{Character scalar. Title of the plot}

\item{...}{Further arguments passed to \code{\link{matplot}}.}
}
\description{
Visualize adopters and cumulative adopters
}
\seealso{
Other visualizations: \code{\link{hazard_rate}},
  \code{\link{plot_diffnet}},
  \code{\link{plot_infectsuscep}},
  \code{\link{plot_threshold}}
}

