% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdiffusr_classes.R
\name{plot_diffnet}
\alias{plot_diffnet}
\title{Create a diffusion graph}
\usage{
plot_diffnet(graph, cumadopt, vcols = c("blue", "grey"),
  mode = "fruchtermanreingold", layout.par = NULL, mfcol.par = NULL, ...)
}
\arguments{
\item{graph}{An array}

\item{cumadopt}{nxT matrix}

\item{vcols}{A vector of size 2 with colors}

\item{mode}{Layout}

\item{layout.par}{Layout parameters}

\item{...}{Further arguments to be passed to gplot}
}
\description{
Create a diffusion graph
}

