% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffnet2.R
\name{plot_diffnet2}
\alias{plot_diffnet2}
\alias{plot_diffnet2.default}
\alias{plot_diffnet2.diffnet}
\title{Another way of visualizing diffusion}
\usage{
plot_diffnet2(...)

\method{plot_diffnet2}{diffnet}(graph, slice = nslices(graph),
  color.ramp = grDevices::colorRamp(c("skyblue", "yellow", "red")),
  layout = NULL, key.width = 0.2, key.title = "Time of Adoption",
  main = "Diffusion dynamics", vertex.size = NULL, vertex.shape = NULL,
  vertex.label = "", vertex.frame.color = "gray", edge.arrow.size = 0.5,
  edge.curved = FALSE, add.map = FALSE, diffmap.args = list(kde2d.args =
  list(n = 100)), include.white = "first", ...)

\method{plot_diffnet2}{default}(graph, toa, pers = min(toa, na.rm =
  TRUE):max(toa, na.rm = TRUE), color.ramp = grDevices::colorRamp(c("skyblue",
  "yellow", "red")), layout = NULL, key.width = 0.2,
  key.title = "Time of Adoption", main = "Diffusion dynamics",
  vertex.size = NULL, vertex.shape = NULL, vertex.label = "",
  vertex.frame.color = "gray", edge.arrow.size = 0.5, edge.curved = FALSE,
  add.map = FALSE, diffmap.args = list(kde2d.args = list(n = 100)),
  include.white = "first", ...)
}
\arguments{
\item{...}{Further arguments passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{graph}{Either a square matrix or a diffnet object.}

\item{color.ramp}{A function as returned by \code{\link[grDevices::colorRamp]{colorRamp}}.}

\item{layout}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{key.width}{Numeric scalar. Sets the proportion of the plot (x-axis) that the key uses.}

\item{key.title}{Character scalar. Title of the key (vertex colors).}

\item{main}{Character scalar. Title of the graph.}

\item{vertex.size}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{vertex.shape}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{vertex.label}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{vertex.frame.color}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{edge.arrow.size}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{edge.curved}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{add.map}{Logical scalar. When \code{TRUE} adds plots \code{\link{diffusionMap}}.}

\item{diffmap.args}{List. If \code{add.map=TRUE}, arguments passed to \code{diffusionMap}.}

\item{include.white}{Character scalar. Includes white in the color palette used in the map.
When \code{include.white=NULL} then it won't include it.}

\item{toa}{Integer vector of length \eqn{n} with the times of adoption.}

\item{pers}{Integer vector of length \eqn{T} indicating the time periods of the data.}
}
\value{
A list with the following elements
\item{layout}{}
\item{vertex.color}{}
\item{vertex.label}{}
\item{vertex.shape}{}
\item{vertex.size}{}
\item{diffmap}{}
}
\description{
Another way of visualizing diffusion
}
\details{
If \code{key.width<=0} then no key is created.
}
\seealso{
Other visualizations: \code{\link{diffusionMap}},
  \code{\link{hazard_rate}}, \code{\link{plot_adopters}},
  \code{\link{plot_diffnet}},
  \code{\link{plot_infectsuscep}},
  \code{\link{plot_threshold}}
}

