% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{random}
\alias{random}
\title{net 19}
\format{\preformatted{'data.frame':	7267 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  1 1 1 1 1 1 1 1 1 1 ...
 $ alter : num  11 18 26 35 36 38 42 44 46 56 ...
 $ dichot: num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
random
}
\description{
net 19
}
\keyword{datasets}

