% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ratify}
\alias{ratify}
\title{Data on ratification of 4 treaties}
\format{\preformatted{'data.frame':	764 obs. of  4 variables:
 $ id           : int  1 1 1 1 2 2 2 2 3 3 ...
 $ countrynameun: chr  "Afghanistan" "Afghanistan" "Afghanistan" "Afghanistan" ...
 $ treaty       : int  1 2 3 4 1 2 3 4 1 2 ...
 $ year         : num  9 7 10 10 5 5 3 5 5 3 ...
}}
\source{
un treaty database
}
\usage{
ratify
}
\description{
Data on ratification of 4 treaties
}
\keyword{datasets}

