% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recode}
\alias{recode}
\alias{recode.data.frame}
\alias{recode.matrix}
\title{Recodes an edgelist such that ids go from 1 to n}
\usage{
recode(data, ...)

\method{recode}{data.frame}(data, ...)

\method{recode}{matrix}(data, ...)
}
\arguments{
\item{data}{Edgelist as either a matrix or dataframe with ego and alter}

\item{...}{Further arguments for the method (ignored)}
}
\value{
A recoded edgelist
}
\description{
Recodes an edgelist such that ids go from 1 to n
}
\details{
Recomended for ease of use
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for data frame

\item \code{matrix}: Method for matrix
}}
\examples{
edgelist <- cbind(c(1,1,3,6),c(4,3,200,1))
edgelist
recode(edgelist)
}

