% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graph.R
\name{rgraph_ws}
\alias{rgraph_ws}
\title{Watts-Strogatz model}
\usage{
rgraph_ws(n, k, p, both.ends = FALSE, self = FALSE, multiple = FALSE)
}
\arguments{
\item{n}{Integer scalar. Set the size of the graph.}

\item{k}{Integer scalar. Set the initial degree of the ring (must be less than \eqn{n}).}

\item{p}{Numeric scalar. Set the probability of changing an edge.}

\item{both.ends}{Logical scalar. When \code{TRUE} rewires both ends.}

\item{self}{Logical scalar. When \code{TRUE}, allows loops (self edges).}

\item{multiple}{Logical scalar. When \code{TRUE} allows multiple edges.}
}
\value{
A random graph of size \eqn{n\times n}{n*n} following the small-world
model.
}
\description{
Watts-Strogatz model
}

