% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{select_egoalter}
\alias{select_egoalter}
\alias{select_egoalter.array}
\title{Selection function}
\usage{
select_egoalter(...)

\method{select_egoalter}{array}(adjmat, adopt, period = NULL)
}
\arguments{
\item{adjmat}{Adjacency matrix}

\item{adopt}{Adoption matrix}

\item{period}{Integer time period}
}
\description{
Selection function
}
\section{Methods (by class)}{
\itemize{
\item \code{array}: Method for array
}}

