% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectionFunctionEgoAltere}
\alias{selectionFunctionEgoAltere}
\title{Calculate the number of times the 16 selection changes occurred.}
\usage{
selectionFunctionEgoAltere(all_nets, Adopt_mat, Time)
}
\arguments{
\item{all_nets}{An array of networks stored as adjacency matrices.}

\item{Adopt_mat}{Matrix of row=nodes, column=year, Binary if adopted by year.}

\item{Time}{Timepoint in which calculating selection for.}
}
\value{
Matrix of of the count of selection changes for the 16 categories by node.
}
\description{
Calculate the number of times the 16 selection changes occurred.
}

