% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_equiv.R
\name{struct_equiv}
\alias{struct_equiv}
\title{Conmputes Structural Equivalence}
\usage{
struct_equiv(adjmat, v = 1, ...)
}
\arguments{
\item{adjmat}{Square matrix. Adjacency matrix}

\item{v}{Double. Constant for SE}

\item{...}{Further arguments to be passed to \code{sna::geodist}}
}
\value{
A square matrix indicating structural equivalence.
}
\description{
Conmputes Structural Equivalence
}

