% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_to_diffnet.R
\name{survey_to_diffnet}
\alias{survey_to_diffnet}
\title{Convert survey-like data to a diffnet object}
\usage{
survey_to_diffnet(dat, idvar, netvars, toavar, groupvar = NULL,
  no.unsurveyed = TRUE, timevar = NULL, t = NULL,
  undirected = getOption("diffnet.undirected", FALSE),
  self = getOption("diffnet.self", FALSE),
  multiple = getOption("diffnet.multiple", FALSE), keep.isolates = TRUE,
  recode.ids = TRUE, ...)
}
\arguments{
\item{dat}{A data frame.}

\item{idvar}{Character scalar. Name of the id variable.}

\item{netvars}{Character vector. Names of the network nomination variables.}

\item{toavar}{Character scalar. Name of the time of adoption variable.}

\item{groupvar}{Character scalar. Name of cohort variable (e.g. city).}

\item{no.unsurveyed}{Logical scalar. When TRUE the nominated individuals
that do not show in \code{idvar} are set to \code{NA} (see details).}

\item{timevar}{Character sacalar. In the case of longitudinal data, name of the time var.}

\item{t}{Integer scalar. Repeat the network \code{t} times (if no \code{t0,t1} are provided).}

\item{undirected}{Logical scalar. TRUE when the graph is undirected.}

\item{self}{Logical scalar. TRUE when self edges are excluded.}

\item{multiple}{Logical scalar. TRUE when multiple edges should not be included
(see details).}

\item{keep.isolates}{Logical scalar. When FALSE, rows with \code{NA/NULL} values
(isolated vertices unless have autolink) will be droped (see details).}

\item{recode.ids}{Logical scalar. When TRUE ids are recoded using \code{\link{as.factor}}
(see details).}

\item{...}{Further arguments to be passed to \code{\link{as_diffnet}}}
}
\value{
A \code{\link{diffnet}} object.
}
\description{
This convenient function turns network nomination data into diffnet obejects.
It works as a wrapper of \code{\link{edgelist_to_adjmat}}
and \code{\link{as_diffnet}}, and as a preprocessing routine of survey-like data,
creating the respective edgelist to be imported as a network.
}
\details{
All of \code{idvar}, \code{netvars}, \code{toavar} and \code{groupvar}
must be integers. Were these numeric they are coerced into integers, otherwise,
when neither of both, the function returns with error.

In field work it is not unusual that some respondents nominate unsurveyed
individuals. In such case, in order to exclude them from the analysis,
the user can set \code{no.unsurveyed=TRUE} (the default), telling the
function to exclude such individuals from the adjacency matrix. This is
done by setting variables in \code{netvars} equal to \code{NA} when the
nominated id can't be found in \code{idvar}.

If the network nomination process was done in different groups (location
for example) the survey id numbers may be define uniquely within each group
but not across groups (there may be many individuals with \code{id=1},
for example). To encompass this issue, the user can tell the function what
variable can be used to distinguish between groups through the \code{groupvar}
argument. When \code{groupvar} is provided, function redifines \code{idvar}
and the variables in \code{netvars} as follows:

\preformatted{
   dat[[idvar]] <- dat[[idvar]] + dat[[groupvar]]*z
}

Where \code{z = 10^nchar(max(dat[[idvar]]))}.

For longitudinal data, it is assumed that the \code{toavar} holds the same
information through time, this is, time-invariable. This as the package does
not yet support variable times of adoption.
}
\examples{
# Loading a fake survey (data frame)
data(fakesurvey)

# Diffnet object keeping isolated vertices ----------------------------------
dn1 <- survey_to_diffnet(fakesurvey, "id", c("net1", "net2", "net3"), "toa",
   "group", keep.isolates=TRUE)

# Diffnet object NOT keeping isolated vertices
dn2 <- survey_to_diffnet(fakesurvey, "id", c("net1", "net2", "net3"), "toa",
   "group", keep.isolates=FALSE)

# dn1 has an extra vertex than dn2
dn1
dn2

# Loading a longitudinal survey data (two waves) ----------------------------
data(fakesurveyDyn)

groupvar <- "group"
x <- survey_to_diffnet(
   fakesurveyDyn, "id", c("net1", "net2", "net3"), "toa", "group" ,
   timevar = "time", keep.isolates = TRUE)

plot_diffnet(x, vertex.cex = 1.5, displaylabels = TRUE)

# Reproducing medInnovationsDiffNet object ----------------------------------
data(medInnovations)

# What are the netvars
netvars <- names(medInnovations)[grepl("^net", names(medInnovations))]

medInnovationsDiffNet2 <- survey_to_diffnet(
   medInnovations,
   "id", netvars, "toa", "city")

medInnovationsDiffNet2

# Comparing with the package's version
all(diffnet.toa(medInnovationsDiffNet2) == diffnet.toa(medInnovationsDiffNet)) #TRUE
all(
   diffnet.attrs(medInnovationsDiffNet2, as.df = TRUE) ==
   diffnet.attrs(medInnovationsDiffNet, as.df = TRUE),
   na.rm=TRUE) #TRUE


}
\author{
Vega Yon
}
\seealso{
\code{\link{fakesurvey}}
}

