% Generated by roxygen2 (4.0.2): do not edit by hand
\name{susceptInfecte}
\alias{susceptInfecte}
\title{Calculates susceptibility and infectiousness}
\usage{
susceptInfecte(toa, toamat, Adopt_mat1, all_nets)
}
\arguments{
\item{toa}{timevector}

\item{toamat}{difference in adoption times between all nodes in network}

\item{Adopt_mat1}{}

\item{all_nets}{adjacency matrices}

\item{maxTime}{number of timepoints}
}
\description{
Calculates susceptibility and infectiousness
}

