% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{threshold}
\alias{threshold}
\title{Calculates threshold}
\usage{
threshold(exposure, toe)
}
\arguments{
\item{exposure}{nxT matrix. Exposure to the innovation obtained from
\code{\link{exposure}}.}

\item{toe}{Integer vector. Indicating the time of adoption of the innovation.}
}
\value{
A vector of size n indicating the threshold of each node.
}
\description{
Calculates threshold
}

