% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{toa_mat}
\alias{toa_mat}
\alias{toa_mat.integer}
\alias{toa_mat.numeric}
\title{Difference in Time of Adoption (TOA) between nodes}
\usage{
toa_mat(times, recode = TRUE, ...)

\method{toa_mat}{integer}(times, recode = TRUE, ...)

\method{toa_mat}{numeric}(times, recode = TRUE, ...)
}
\arguments{
\item{times}{Integer vector containing time of adoption of the innovation.}

\item{recode}{Logical. When TRUE recodes time (see details).}

\item{...}{Ignored.}
}
\value{
An n x n matrix indicating the difference in times of adoption between
each pair of nodes.
}
\description{
Creates n x n matrix indicating the difference in times of adoption between
each pair of nodes
}
\section{Methods (by class)}{
\itemize{
\item \code{integer}: Integer vectors

\item \code{numeric}: Numeric vectors
}}
\examples{
# Generating a random vector of time
set.seed(123)
times <- sample(2000:2005, 10, TRUE)

# Computing the TOA differences
toa_mat(times)
}

