% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toaMate}
\alias{toaMate}
\title{Creates #nodes x #nodes matrix indicating the difference in times of adoption between each pair of nodes}
\usage{
toaMate(timevec, id)
}
\arguments{
\item{timevec}{Vector indicating the timepoint in which each node adopted the innovation}
}
\value{
#nodes x #nodes matrix indicating the difference in times of adoption between each pair of nodes
}
\description{
Creates #nodes x #nodes matrix indicating the difference in times of adoption between each pair of nodes
}

