% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tobtrade}
\alias{tobtrade}
\title{net 3}
\format{\preformatted{'data.frame':	43494 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  2 2 2 2 2 3 3 3 3 3 ...
 $ alter : num  67 98 120 150 165 31 61 76 98 112 ...
 $ dichot: num  1 0 1 1 0 0 0 0 0 0 ...
}}
\usage{
tobtrade
}
\description{
net 3
}
\keyword{datasets}

