% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{trade}
\alias{trade}
\title{net 2}
\format{\preformatted{'data.frame':	181612 obs. of  4 variables:
 $ year  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ id    : num  2 2 2 2 2 2 2 2 2 2 ...
 $ alter : num  1 10 17 22 26 31 36 37 41 43 ...
 $ dichot: num  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
trade
}
\description{
net 2
}
\keyword{datasets}

