% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertex_covariate_dist}
\alias{vertex_covariate_dist}
\title{Computes p-norm between connected vertices}
\usage{
vertex_covariate_dist(graph, X, p = 2)
}
\arguments{
\item{graph}{A square matrix of size \eqn{n} of class dgCMatrix.}

\item{X}{A numeric matrix of size \eqn{n\times K}{n * K}. Vertices attributes}

\item{p}{Numeric scalar. Norm to compute}
}
\value{
A matrix of size \eqn{n\times n}{n*n} of class \code{dgCMatrix}. Will
be symmetric only if \code{graph} is symmetric.
}
\description{
Computes p-norm between connected vertices
}
\details{
For each par of vertices, the function computes the following
\deqn{%
D_{ij} = \left(\sum_{k=1}^K (X_{ik} - X_{jk})^{p} \right)^{1/p}\mbox{ if }graph_{i,j}\neq 0
}{%
D(i,j) = [\sum_k (X(i,k) - X(j,k))^p]^(1/p)  if graph(i,j) != 0
}
}
\examples{
set.seed(123)
G <- rgraph_ws(20, 4, .1)
X <- matrix(runif(40), ncol=2)

vertex_covariate_dist(G, X)
}

