% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{wku_diffnet}
\alias{wku_diffnet}
\title{WKU Diffusion Network}
\format{
A \code{\link{diffnet}} class object.
}
\source{
WKU simulation
}
\description{
A \code{\link{diffnet}} object with 594 vertices and 15 time periods from the
WKU simulation study.
}
\seealso{
Other diffusion datasets: 
\code{\link{brfarmers}},
\code{\link{brfarmersDiffNet}},
\code{\link{diffusion-data}},
\code{\link{epigames}},
\code{\link{epigamesDiffNet}},
\code{\link{fakeDynEdgelist}},
\code{\link{fakeEdgelist}},
\code{\link{fakesurvey}},
\code{\link{fakesurveyDyn}},
\code{\link{kfamily}},
\code{\link{kfamilyDiffNet}},
\code{\link{medInnovations}},
\code{\link{medInnovationsDiffNet}}
}
\concept{diffusion datasets}
