% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdiffusr_classes.R
\name{as_diffnet}
\alias{as_diffnet}
\title{Creates a \code{diffnet} class object}
\usage{
as_diffnet(data, toa, weights = NULL, times = NULL, undirected = FALSE,
  self = FALSE, multiple = FALSE, use.incomplete = FALSE, recode = TRUE,
  ...)
}
\arguments{
\item{data}{Either an adjacency matrix, an array or an edgelist}

\item{recode}{Logical scalar. Passed to \code{\link{toa_mat}}}

\item{tao}{Numeric vector of size \eqn{n}. Times of adoption.}
}
\description{
Creates a \code{diffnet} class object
}

