% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_diffnet}
\alias{as_diffnet}
\alias{print.diffnet}
\alias{summary.diffnet}
\title{Creates a \code{diffnet} class object}
\usage{
as_diffnet(graph, toa, recode = TRUE, weights = NULL, times = NULL,
  undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"),
  multiple = getOption("diffnet.multiple"), use.incomplete = FALSE, ...)

\method{print}{diffnet}(x, ...)

\method{summary}{diffnet}(object, ...)
}
\arguments{
\item{graph}{A dynamic graph (see \code{\link{netdiffuseR-graphs}}).}

\item{toa}{Numeric vector of size \eqn{n}. Times of adoption.}

\item{recode}{Logical scalar. Passed to \code{\link{toa_mat}}.}

\item{weights}{Numeric vector of size \eqn{n}.}

\item{times}{Numeric vector of size \eqn{n}.}

\item{undirected}{Logical scalar.}

\item{self}{Logical scalar.}

\item{multiple}{Logical scalar.}

\item{use.incomplete}{Logical scalar.}

\item{...}{Ignored.}
}
\description{
Creates a \code{diffnet} class object
}
\examples{
# Creating a random graph
set.seed(123)
graph <- rgraph_ba(t=9)
graph <- lapply(1:5, function(x) graph)

# Pretty TOA
names(graph) <- 2001L:2005L
toa <- sample(c(2001L:2005L,NA), 10, TRUE)

# Creating diffnet object
diffnet <- as_diffnet(graph, toa)
diffnet
summary(diffnet)
}
\seealso{
Default options are listed at \code{\link{netdiffuseR-options}}
}

