% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_diffnet}
\alias{as_diffnet}
\alias{diffnet}
\alias{diffnet-class}
\alias{plot.diffnet}
\alias{print.diffnet}
\alias{summary.diffnet}
\alias{toa}
\alias{toa<-}
\title{Creates a \code{diffnet} class object}
\usage{
as_diffnet(graph, toa, t0 = min(toa, na.rm = TRUE), t1 = max(toa, na.rm =
  TRUE), weights = NULL, undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"),
  multiple = getOption("diffnet.multiple"), use.incomplete = FALSE)

toa(graph)

toa(graph, i) <- value

\method{plot}{diffnet}(x, y = NULL, t = 1, displaylabels = FALSE,
  vertex.col = c("blue", "grey"), gmode = ifelse(x$meta$undirected, "graph",
  "digraph"), vertex.cex = "degree", edge.col = "gray",
  mode = "fruchtermanreingold", layout.par = NULL,
  main = "Diffusion network in time \%d", ...)

\method{print}{diffnet}(x, ...)

\method{summary}{diffnet}(object, ...)
}
\arguments{
\item{graph}{A dynamic graph (see \code{\link{netdiffuseR-graphs}}).}

\item{toa}{Numeric vector of size \eqn{n}. Times of adoption.}

\item{t0}{Integer scalar. Passed to \code{\link{toa_mat}}.}

\item{t1}{Integer scalar. Passed to \code{\link{toa_mat}}.}

\item{weights}{Numeric vector of size \eqn{n}.}

\item{undirected}{Logical scalar.}

\item{self}{Logical scalar.}

\item{multiple}{Logical scalar.}

\item{use.incomplete}{Logical scalar.}

\item{i}{Indices specifying elements to replace. See \code{\link[base:Extract]{Extract}}.}

\item{value}{Replacement.}

\item{x}{A \code{diffnet} object.}

\item{y}{Ignored.}

\item{t}{Integer scalar indicating the time slice to plot.}

\item{displaylabels}{Logical scalar. When TRUE, \code{plot} shows vertex labels.}

\item{vertex.col}{Character scalar/vector. Color of the vertices.}

\item{gmode}{Character scalar. Passed to \code{\link[sna:gplot]{gplot}.}}

\item{vertex.cex}{Numeric scalar/vector. Size of the vertices.}

\item{edge.col}{Character scalar/vector. Color of the edges.}

\item{mode}{Character scalar. Name of the layout algorithm to implement (see details).}

\item{layout.par}{Layout parameters (see details).}

\item{main}{Character. A title template to be passed to sprintf.}

\item{...}{In the case of \code{plot}, further arguments passed to \code{gplot}, otherwise
is ignored.}

\item{object}{A \code{diffnet} object.}
}
\value{
A list of class \code{diffnet} with the following elements:
\item{graph}{A list of length \eqn{T}. Containing sparse square matrices of size \eqn{n}
and class \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}.}
\item{toa}{An integer vector of size \eqn{T} with times of adoption.}
\item{adopt, cumadopt}{Numeric matrices of size \eqn{n\times T}{n*T} as those returned
by \code{\link{toa_mat}}.}
\item{meta}{A list of length 9 with the following elements:
\itemize{
 \item \code{type}: Character scalar equal to \code{"dynamic"}.
 \item \code{class}: Character scalar equal to \code{"list"}.
 \item \code{ids}: Character vector of size \eqn{n} with vertices' labels.
 \item \code{pers}: Integer vector of size \eqn{T}.
 \item \code{nper}: Integer scalar equal to \eqn{T}.
 \item \code{n}: Integer scalar equal to \eqn{n}.
 \item \code{self}: Logical scalar.
 \item \code{undirected}: Logical scalar.
 \item \code{multiple}: Logical scalar.
}
}
}
\description{
\code{diffnet} objects contain difussion of innovation networks. With adjacency
matrices and time of adoption (toa) vector as its main components, most of the
package's functions have methods for this class of objects.
}
\details{
Plotting is done via the function \code{\link[sna:gplot]{gplot}},
and its layout via \code{\link[sna:gplot.layout]{gplot.layout}}, both from
the (\pkg{sna}) package.

\code{vertex.cex} can either be a numeric scalar, a numeric vector or a character
scalar taking any of the following values \code{"degree"}, \code{"indegree"}, or
\code{"outdegree"}. The later will be passed to \code{\link{dgr}} to calculate
degree of the selected slice and will be normalized as

\deqn{vertex.cex = d/[max(d) - min(d)]\times 2 + .5}{vertex.cex = d/[max(d) - min(d)]* 2 + .5}

where \code{d=sqrt(dgr(graph))}.

In the case of the \code{summary} method, Moran's I is calculated over the
cumulative adoption matrix using as weighting matrix the inverse of the geodesic
distance matrix. All this via \code{\link{moran}}.
}
\section{Auxiliary functions}{


\code{toa(graph)} Works as an alias of \code{graph$toa}. While this function
may not be very useful, the replacement method, \code{toa<-} used as \code{toa(graph)<-...},
is the right way of modifying times of adoption as when the \code{toa} vector
is changed the method performs several checks on the time ranges, and
recalculates adoption and cumulative adoption matrices using \code{toa_mat}.
#'
}
\examples{

# Creating a random graph
set.seed(123)
graph <- rgraph_ba(t=9)
graph <- lapply(1:5, function(x) graph)

# Pretty TOA
names(graph) <- 2001L:2005L
toa <- sample(c(2001L:2005L,NA), 10, TRUE)

# Creating diffnet object
diffnet <- as_diffnet(graph, toa)
diffnet
summary(diffnet)
}
\seealso{
Default options are listed at \code{\link{netdiffuseR-options}}
}

