% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_diffnet}
\alias{as_diffnet}
\alias{diffnet}
\alias{diffnet-class}
\alias{plot.diffnet}
\alias{print.diffnet}
\alias{summary.diffnet}
\title{Creates a \code{diffnet} class object}
\usage{
as_diffnet(graph, toa, recode = TRUE, weights = NULL,
  undirected = getOption("diffnet.undirected"),
  self = getOption("diffnet.self"),
  multiple = getOption("diffnet.multiple"), use.incomplete = FALSE, ...)

\method{plot}{diffnet}(x, y = NULL, t = 1, displaylabels = FALSE,
  vertex.col = c("blue", "grey"), vertex.cex = NA, edge.col = "gray",
  mode = "fruchtermanreingold", layout.par = NULL,
  main = "Diffusion network in time \%d", ...)

\method{print}{diffnet}(x, ...)

\method{summary}{diffnet}(object, ...)
}
\arguments{
\item{graph}{A dynamic graph (see \code{\link{netdiffuseR-graphs}}).}

\item{toa}{Numeric vector of size \eqn{n}. Times of adoption.}

\item{recode}{Logical scalar. Passed to \code{\link{toa_mat}}.}

\item{weights}{Numeric vector of size \eqn{n}.}

\item{undirected}{Logical scalar.}

\item{self}{Logical scalar.}

\item{multiple}{Logical scalar.}

\item{use.incomplete}{Logical scalar.}

\item{...}{In the case of \code{plot}, further arguments passed to \code{gplot}, otherwise
is ignored.}

\item{x}{A \code{diffnet} object.}

\item{y}{Ignored.}

\item{t}{Integer scalar indicating the time slice to plot.}

\item{displaylabels}{Logical scalar. When TRUE, \code{plot} shows vertex labels.}

\item{vertex.col}{Character scalar/vector. Color of the vertices.}

\item{vertex.cex}{Numeric scalar/vector. Size of the vertices.}

\item{edge.col}{Character scalar/vector. Color of the edges.}

\item{mode}{Character scalar. Name of the layout algorithm to implement (see details).}

\item{layout.par}{Layout parameters (see details).}

\item{main}{Character. A title template to be passed to sprintf.}

\item{object}{A \code{diffnet} object.}
}
\description{
Creates a \code{diffnet} class object
}
\details{
Plotting is done via the function \code{\link[sna:gplot]{gplot}},
and its layout via \code{\link[sna:gplot.layout]{gplot.layout}}, both from
the (\pkg{sna}) package.
}
\examples{

# Creating a random graph
set.seed(123)
graph <- rgraph_ba(t=9)
graph <- lapply(1:5, function(x) graph)

# Pretty TOA
names(graph) <- 2001L:2005L
toa <- sample(c(2001L:2005L,NA), 10, TRUE)

# Creating diffnet object
diffnet <- as_diffnet(graph, toa)
diffnet
summary(diffnet)
}
\seealso{
Default options are listed at \code{\link{netdiffuseR-options}}
}

