% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{degree}
\alias{degree}
\alias{degree.array}
\alias{degree.matrix}
\title{Indegree, outdegree and degree of the vertices}
\usage{
degree(graph, cmode = "degree", undirected = TRUE, self = FALSE)

\method{degree}{matrix}(graph, cmode = "degree", undirected = TRUE,
  self = FALSE)

\method{degree}{array}(graph, cmode = "degree", undirected = TRUE,
  self = FALSE)
}
\arguments{
\item{graph}{Either an \eqn{n\times n}{n * n} Matrix or a \eqn{n\times n\times T}{n * n * T} array}

\item{cmode}{Character. Either "indegree", "outdegree" or "degree".}

\item{undirected}{Logical. TRUE when the graph is undirected.}

\item{self}{Logical. TRUE when self edges should not be considered.}
}
\value{
Either a numeric vector of size \eqn{n}{n} with the degree of each node (if graph is
a matrix), or a matrix of size \eqn{n\times T}{n * T}
}
\description{
Computes the requested degree measure for each node in the graph.
}
\examples{
# Creating a directed graph
graph <- rand_graph(undirected=FALSE)
graph

# Comparing degree measurements
 data.frame(
   In=degree(graph, 0, undirected = FALSE),
   Out=degree(graph, 1, undirected = FALSE),
   Degree=degree(graph, 2, undirected = FALSE)
 )
}

