% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egonets.R
\name{egonet_attrs}
\alias{egonet_attrs}
\title{Retrieve alter's attributes}
\usage{
egonet_attrs(graph, attrs, E = NULL, outer = TRUE,
  self = getOption("diffnet.self"), valued = getOption("diffnet.valued"))
}
\arguments{
\item{graph}{Any kind of graph}
}
\description{
Retrieve alter's attributes
}

