% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{epigamesDiffNet}
\alias{epigamesDiffNet}
\title{Epi Games Dataset (DiffNet version)}
\format{
A `diffnet` object. See `[diffnet-class]` for more details.
}
\source{
WKU Epi Game simulation
}
\description{
The `diffnet` version of `[epigames_raw]`. Note that following `netdiffuseR` 
conventions, right-censored non-adopters (`status == "not_infected"`) have their 
`toa` values set to 16 (`max(time) + 1`) within the `diffnet` object.
}
\seealso{
Other diffusion datasets: 
\code{\link{brfarmers}},
\code{\link{brfarmersDiffNet}},
\code{\link{diffusion-data}},
\code{\link{epigames_raw}},
\code{\link{fakeDynEdgelist}},
\code{\link{fakeEdgelist}},
\code{\link{fakesurvey}},
\code{\link{fakesurveyDyn}},
\code{\link{kfamily}},
\code{\link{kfamilyDiffNet}},
\code{\link{medInnovations}},
\code{\link{medInnovationsDiffNet}}
}
\concept{diffusion datasets}
