% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{epigamesDiffNet}
\alias{epigamesDiffNet}
\title{\code{diffnet} version of the Epi Games data}
\format{
A \code{\link{diffnet}} class object.
}
\source{
WKU Epi Game simulation
}
\description{
A directed dynamic graph with 594 vertices and 15 time periods. The attributes
in the graph are described in \code{\link{epigames}}.
}
\details{
Note that following \code{netdiffuseR} conventions, right-censored non-adopters
have their \code{toa} values set to 16 (\code{max(time) + 1}).
}
\seealso{
Other diffusion datasets: 
\code{\link{brfarmers}},
\code{\link{brfarmersDiffNet}},
\code{\link{diffusion-data}},
\code{\link{epigames}},
\code{\link{fakeDynEdgelist}},
\code{\link{fakeEdgelist}},
\code{\link{fakesurvey}},
\code{\link{fakesurveyDyn}},
\code{\link{kfamily}},
\code{\link{kfamilyDiffNet}},
\code{\link{medInnovations}},
\code{\link{medInnovationsDiffNet}},
\code{\link{wku_diffnet}}
}
\concept{diffusion datasets}
