% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{exposure}
\alias{exposure}
\title{Ego exposure}
\usage{
exposure(graph, cumadopt, wtype = 0, v = 1,
  undirected = getOption("diffnet.undirected"), normalized = TRUE)
}
\arguments{
\item{graph}{A dynamic graph (see \code{\link{netdiffuseR-graphs}}).}

\item{cumadopt}{nxT matrix. Cumulative adoption matrix obtained from
\code{\link{toa_mat}}}

\item{wtype}{Integer. Weighting type (see details).}

\item{v}{Double. Constant for Structural Equivalence.}

\item{undirected}{Logical. TRUE if the graph is undirected.}

\item{normalized}{Logical. When true, the exposure will be between zero
and one (see details).}
}
\value{
A matrix of size nxT with exposure for each node.
}
\description{
Calculates the level of exposure to the innovation that each node has in the
graph.
}
\details{
When \code{wtype=0} (default), exposure is defined as follows

\deqn{E_{n(t)}=\frac{S_{nn}\times A_{n(t)}}{S_{n+}}}{%
      E(n,t)  =[     S(n,n) x     A(n,t)] / S(n+)}

where \eqn{E_{n(t)}}{E(n,t)} is the exposure of the network at time t,
\eqn{S_{nn}}{S(n,n)} is the social network, \eqn{A_{n(t)}}{A(n,t)} is the
cumulative adoption matrix at time t, and \eqn{S_{n+}}{S(n+)} is the row-wide
sum of the graph.

If \code{wtype=1} (Structural Equivalence), exposure is now computed as:

\deqn{E_{n(t)}=\frac{SE_{nn}^{-1}\times A_{n(t)}}{SE_{n+}^{-1}}}{%
      E(n,t)  =[     SE(n,n)^[-1] x     A(n,t)] / SE(n+)^[-1]}

where SE stands for Structural Equivalence (see \code{\link{struct_equiv}}).
Otherwise, for any value above of \code{wtype}--2, 3 or 4, which stands for
indegree, outdegree and degree respectively-- is calculated accordingly to

\deqn{E_{n(t)}=\frac{S_{nn}\times (A_{n(t)} / D_n)}{S_{n+}}}{%
      E(n,t)  =[     S(n,n) x     [A(n,t)/D(n)] / S(n+)}

where \eqn{D_n}{D(n)} is a column vector of size n containing the degree of
each node.

Finally, note that whenever \code{normalized=TRUE}, the resulting output is only the
numerator of the above formulas.
}
\references{
Burt, R. S. (1987). "Social Contagion and Innovation: Cohesion versus Structural
Equivalence". American Journal of Sociology, 92(6), 1287.
\url{http://doi.org/10.1086/228667}

Valente, T. W. (1995). "Network models of the diffusion of innovations"
 (2nd ed.). Cresskill N.J.: Hampton Press.
}
\seealso{
Other statistics: \code{\link{cumulative_adopt_count}},
  \code{\link{dgr}}, \code{\link{hazard_rate}},
  \code{\link{infection}}, \code{\link{struct_equiv}},
  \code{\link{threshold}}
}
\keyword{univar}

