% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fakeDynEdgelist}
\alias{fakeDynEdgelist}
\title{Fake dynamic edgelist}
\format{A data frame with 22 rows and 4 variables
\describe{
 \item{ego}{Nominating individual}
 \item{alter}{Nominated individual}
 \item{value}{Strength of the tie}
 \item{time}{Integer with the time of the spell}
}}
\source{
Generated for the package
}
\usage{
fakeDynEdgelist
}
\description{
A data frame used for examples in reading edgelist format networks. This
edgelist can be merged with the dataset \code{\link{fakesurveyDyn}}.
}
\author{
George G. Vega Yon
}
\seealso{
Other diffusion datasets: \code{\link{brfarmersDiffNet}},
  \code{\link{brfarmers}}, \code{\link{diffusion-data}},
  \code{\link{fakeEdgelist}}, \code{\link{fakesurveyDyn}},
  \code{\link{fakesurvey}}, \code{\link{kfamilyDiffNet}},
  \code{\link{kfamily}},
  \code{\link{medInnovationsDiffNet}},
  \code{\link{medInnovations}}
}
\keyword{datasets}

