% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fakeEdgelist}
\alias{fakeEdgelist}
\title{Fake static edgelist}
\format{A data frame with 11 rows and 3 variables
\describe{
 \item{ego}{Nominating individual}
 \item{alter}{Nominated individual}
 \item{value}{Strength of the tie}
}}
\source{
Generated for the package
}
\description{
A data frame used for examples in reading edgelist format networks. This
edgelist can be merged with the dataset \code{\link{fakesurvey}}.
}
\author{
George G. Vega Yon
}
\seealso{
Other diffusion datasets: \code{\link{brfarmersDiffNet}},
  \code{\link{brfarmers}}, \code{\link{diffusion-data}},
  \code{\link{fakeDynEdgelist}},
  \code{\link{fakesurveyDyn}}, \code{\link{fakesurvey}},
  \code{\link{kfamilyDiffNet}}, \code{\link{kfamily}},
  \code{\link{medInnovationsDiffNet}},
  \code{\link{medInnovations}}
}

