% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hazard_rate}
\alias{hazard_rate}
\title{Graph Hazard Rate}
\usage{
hazard_rate(cumadopt)
}
\arguments{
\item{cumadopt}{A \eqn{n\times T}{n * T}. Cumulative adoption matrix obtained from
\code{\link{toa_mat}}}
}
\value{
A row vector of size \eqn{T} with hazard rates for \eqn{t>1}.
}
\description{
Computes the hazard rate of the network at each period of time.
}
\details{
For \eqn{t>1}, hazard rate is calculated as

\deqn{\frac{q_t - q_{t-1}}{n - q_{t-1}}}{[q(t) - q(t-1)]/[n - q(t-1)]}

where \eqn{q_i}{q(i)} is the number of adopters in time \eqn{t}, and \eqn{n} is
the number of vertices in the graph.
}
\seealso{
Other statistics: \code{\link{cumulative_adopt_count}},
  \code{\link{dgr}}, \code{\link{exposure}},
  \code{\link{infection}}, \code{\link{struct_equiv}},
  \code{\link{threshold}}
}
\keyword{univar}

