% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hazard_rate}
\alias{hazard_rate}
\title{Calculates hazard rate}
\usage{
hazard_rate(cumadopt)
}
\arguments{
\item{cumadopt}{nxT matrix. Cumulative adoption matrix obtained from
\code{\link{toa_mat}}}
}
\value{
A row vector of size T with hazard rates for t>1.
}
\description{
Calculates hazard rate
}
\details{
Hazard rate is calculated as
\deqn{\fraq{q_t - q_{t-1}}{n - q_{t-1}}}{[q(t) - q(t-1)]/[n - q(t-1)]}
where \eqn{q_i}{q(i)} is the number of adopters in time i, and \eqn{n}{n} is the number of individuals
in the system.
}

