% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{isolated}
\alias{isolated}
\title{Finds isolated nodes}
\usage{
isolated(adjmat, undirected = TRUE)
}
\arguments{
\item{adjmat}{Square matrix. An graph as an adjacency matrix.}

\item{undirected}{Logical. TRUE when the graph is undirected.}
}
\value{
An integer vector of size n with 1's where a node is isolated
}
\description{
Finds isolated nodes
}

