% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{isolated}
\alias{drop_isolated}
\alias{isolated}
\title{Manages isolated nodes}
\usage{
isolated(adjmat, undirected = TRUE)

drop_isolated(adjmat, undirected = TRUE)
}
\arguments{
\item{adjmat}{Square matrix. An graph as an adjacency matrix.}

\item{undirected}{Logical. TRUE when the graph is undirected.}
}
\value{
In the case of \code{isolated}, an integer vector of size n with 1's where a
node is isolated. Otherwise a modified adjacency matrix excluding the
isolated nodes.
}
\description{
Manages isolated nodes
}
\examples{
\dontrun{
# Generating random graph
set.seed(123)
adjmat <- rand_graph()

# Making nodes 1 and 4 isolated
adjmat[c(1,4),] <- 0
adjmat[,c(1,4)] <- 0

# Finding isolated nodes
isolated(adjmat)

# Removing isolated nodes
drop_isolated(adjmat)
}
}

