% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{isolated}
\alias{drop_isolated}
\alias{drop_isolated.array}
\alias{drop_isolated.matrix}
\alias{isolated}
\alias{isolated.array}
\alias{isolated.matrix}
\title{Find and remove isolated vertices}
\usage{
isolated(graph, undirected = TRUE)

\method{isolated}{matrix}(graph, undirected = TRUE)

\method{isolated}{array}(graph, undirected = TRUE)

drop_isolated(graph, undirected = TRUE)

\method{drop_isolated}{matrix}(graph, undirected = TRUE)

\method{drop_isolated}{array}(graph, undirected = TRUE)
}
\arguments{
\item{graph}{Either a \eqn{n\times n}{n*n} Matrix or a \eqn{n\times n\times T}{n*n*T} array.}

\item{undirected}{Logical. TRUE when the graph is undirected.}
}
\value{
\item{isolated}{an integer vector of size \eqn{n} with 1's where a node is isolated}
 \item{drop_isolated}{a modified graph excluding isolated vertices.}
}
\description{
Find and remove unconnected vertices from the graph.
}
\examples{
\dontrun{
# Generating random graph
set.seed(123)
adjmat <- rand_graph()

# Making nodes 1 and 4 isolated
adjmat[c(1,4),] <- 0
adjmat[,c(1,4)] <- 0

# Finding isolated nodes
isolated(adjmat)

# Removing isolated nodes
drop_isolated(adjmat)
}
}

