% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.r
\docType{package}
\name{netdiffuseR}
\alias{netdiffuseR}
\title{netdiffuseR}
\description{
Statistical analysis, visualization and simulation of diffusion and contagion
processes on networks. The package implements algorithms for
calculating stats such as innovation threshold levels, infectiousness
(contagion) and susceptibility, and hazard rates as presented in Burt (1987),
Valente (1995), and Myers (2000) (among others).

You can access to the project website at
\url{https://github.com/USCCANA/netdiffuseR}
}
\details{
Analysis of Diffusion and Contagion Processes on Networks
}
\section{Acknowledgements}{

netdiffuseR was created with the support of grant R01 CA157577 from the National Cancer Institute/National Institutes of Health.
}

\section{Workshops and Tutorials}{


Online you can find several learning resources:

\itemize{
\item Sunbelt 2016 \url{https://github.com/USCCANA/netdiffuser-sunbelt2016}
\item NASN 2017 \url{https://github.com/USCCANA/netdiffuser-nasn2017}
\item Sunbelt 2018 \url{https://github.com/USCCANA/netdiffuser-sunbelt2018}
}
}

\author{
George G. Vega Yon & Thomas W. Valente
}
