% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffnet-methods.r
\name{plot_diffnet}
\alias{plot_diffnet}
\alias{plot_diffnet.diffnet}
\alias{plot_diffnet.default}
\title{Plot the diffusion process}
\usage{
plot_diffnet(...)

\method{plot_diffnet}{diffnet}(graph, ...)

\method{plot_diffnet}{default}(graph, cumadopt, slices = 1:nslices(graph),
  vertex.col = c("white", "tomato", "steelblue"), vertex.shape = c("square",
  "circle", "circle"), vertex.cex = rowMeans(dgr(graph)), mfrow.par = NULL,
  main = "Network in period \%s", legend.args = list(x = "bottom", legend =
  c("Non adopters", "New adopters", "Adopters"), pch = 21, bty = "n", cex = 1.2,
  horiz = TRUE), intra.space = c(0.15, 0.15), rescale.fun = function(x)
  rescale_vertex_igraph(x), ...)
}
\arguments{
\item{...}{Further arguments to be passed to \code{\link[igraph:plot.igraph]{plot.igraph}}.}

\item{graph}{A dynamic graph (see \code{\link{netdiffuseR-graphs}}).}

\item{cumadopt}{\eqn{n\times T}{n*T} matrix.}

\item{slices}{Integer vector. Indicates what slices to plot. By default all are plotted.}

\item{vertex.col}{A character vector of size 3 with colors names.}

\item{vertex.shape}{A character vector of size 3 with shape names.}

\item{vertex.cex}{Numeric vector of size \eqn{n}. Size of the vertices.}

\item{mfrow.par}{Vector of size 2 with number of rows and columns to be passed to \code{\link{par}.}}

\item{main}{Character scalar. A title template to be passed to \code{\link{sprintf}.}}

\item{intra.space}{Passed to \code{\link[igraph:plot.igraph]{plot.igraph}}}

\item{rescale.fun}{A function to rescale vertex size. By defult it is set to be \code{\link{rescale_vertex_igraph}}}

\item{undirected}{Logical scalar. When \code{TRUE} only the lower triangle of the adjacency matrix will considered (faster).}

\item{lgd}{List of arguments to be passed to \code{\link{legend}}.}
}
\value{
Calculated coordinates for the grouped graph (invisible).
}
\description{
Creates a colored network plot showing the structure of the graph through time
(one network plot for each time period)  and the set of adopter and non-adopters
in the network.
}
\details{
Plotting is done via the function \code{\link[igraph:plot.igraph]{plot.igraph}}.

In order to center the attention on the diffusion process itself, the
positions of each vertex are computed only once by aggregating the networks
through time, this is, instead of computing the layout for each time \eqn{t},
the function creates a new graph accumulating links through time.

The \code{mfrow.par} sets how to arrange the plots on the device. If \eqn{T=5}
and \code{mfrow.par=c(2,3)}, the first three networks will be in the top
of the device and the last two in the bottom.

The argument \code{vertex.col} contains the colors of non-adopters, new-adopters,
and adopters respectively. The new adopters (default color \code{"tomato"}) have a different
color that the adopters when the graph is at their time of adoption, hence,
when the graph been plotted is in \eqn{t=2} and \eqn{toa=2} the vertex will
be plotted in red.
}
\examples{
# Generating a random graph
set.seed(1234)
n <- 6
nper <- 5
graph <- rgraph_er(n,nper, p=.3, undirected = FALSE)
toa <- sample(2000:(2000+nper-1), n, TRUE)
adopt <- toa_mat(toa)

plot_diffnet(graph, adopt$cumadopt)
}
\seealso{
Other visualizations: \code{\link{dgr}},
  \code{\link{diffusionMap}}, \code{\link{drawColorKey}},
  \code{\link{grid_distribution}},
  \code{\link{hazard_rate}}, \code{\link{plot_adopters}},
  \code{\link{plot_diffnet2}},
  \code{\link{plot_infectsuscep}},
  \code{\link{plot_threshold}},
  \code{\link{rescale_vertex_igraph}}
}
\author{
George G. Vega Yon
}
\keyword{hplot}
