% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdiffusr_classes.R
\name{plot_diffnet}
\alias{plot_diffnet}
\title{Plot the diffusion process}
\usage{
plot_diffnet(graph, cumadopt, displaylabels = FALSE, vertex.col = c("blue",
  "grey"), vertex.cex = NA, label = rownames(graph), edge.col = "gray",
  mode = "fruchtermanreingold", layout.par = NULL, mfrow.par = NULL,
  main = "Network in time \%d", mai = c(0, 0, 1, 0), mar = rep(1, 4) +
  0.1, ...)
}
\arguments{
\item{graph}{An array}

\item{cumadopt}{\eqn{n\times T}{n*T} matrix}

\item{displaylabels}{Logical. When TRUE vertex labels are displayed (see \code{\link[sna:gplot]{gplot}})}

\item{vertex.col}{A character vector of size 2 with colors}

\item{vertex.cex}{Numeric vector of size \eqn{n}. Size of the vertices}

\item{label}{Character vector of size \eqn{n}. If no provided, rownames of
the graph are used.}

\item{edge.col}{Character. Color of the edge}

\item{mode}{Character. Name of the layout algorithm to implement (see details)}

\item{layout.par}{Layout parameters (see details)}

\item{mfrow.par}{Vector of size 2 with number of rows and columns to be passed to \code{\link{par}}}

\item{main}{Characetr. A title template to be passed to \code{\link{sprintf}}}

\item{mai}{Numeric vector of size 4. To be passed to \code{\link{par}}}

\item{mar}{Numeric vector of size 4. To be passed to \code{\link{par}}}

\item{...}{Further arguments to be passed to \code{\link[sna:gplot]{gplot}}}
}
\value{
Calculated coordinates (invisible).
}
\description{
Creates a colored network plot showing the structure of the graph through time
(one network plot for each time period)  and the set of adopter and non-adopters
in the network.
}
\details{
Plotting is done via the function \code{\link[sna:gplot]{gplot}},
and its layout via \code{\link[sna:gplot.layout]{gplot.layout}}, both from
the (\pkg{sna}) package.

In order to center the attention on the diffusion process itself, the
positions of each vertex are computed only once by aggregating the networks
through time, this is, instead of computing the layout for each time \eqn{t},
the function creates a new graph accumulating links through time.

The \code{mfrow.par} sets how to arrange the plots on the device. If \eqn{T=5}
and \code{mfrow.par=c(2,3)}, the first three networks will be in the top
of the device and the last two in the bottom.
}
\examples{
#' # Generating a random graph
set.seed(1234)
n <- 6
nper <- 5
graph <- rand_graph(n,nper, p=.3, undirected = FALSE)
toa <- sample(2000:(2000+nper-1), n, TRUE)
adopt <- toa_mat(toa)

plot_diffnet(graph, adopt$cumadopt)
}
\seealso{
Other visualizations: \code{\link{plot_infectsuscep}},
  \code{\link{plot_threshold}}
}
\keyword{hplot}

