% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdiffusr_classes.R
\name{plot_infectsuscep}
\alias{plot_infectsuscep}
\title{Generates a heatmap of \% between suscept and infect}
\usage{
plot_infectsuscep(graph, toa, normalize = TRUE, max.blocks = 100, ...)
}
\arguments{
\item{graph}{an array}

\item{toa}{Times of adoption}

\item{normalize}{Logical. TRUE}

\item{max.blocks}{Maximum number of cols/rows the output matrix will have}

\item{...}{Additional parameters to be passed to infection/susceptibility}
}
\value{
The matrix to be plotted
}
\description{
Generates a heatmap of \% between suscept and infect
}
\examples{
# Generating a random graph
set.seed(1234)
n <- 500
nper <- 20
graph <- rand_graph(n,nper, p=.1, undirected = FALSE)
toa <- sample(1:(1+nper-1), n, TRUE)
adopt <- toa_mat(toa)

x <- plot_infectsuscep(graph, toa, K=3)
image(x, col = heat.colors(100))
}

