% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graph.R
\name{rand_graph}
\alias{rand_graph}
\title{Erdos-Renyi (bernoulli) random graph}
\usage{
rand_graph(n = 10, t = 1, p = 0.3, undirected = TRUE,
  weighted = FALSE, self = FALSE)
}
\arguments{
\item{n}{Integer. Number of vertices}

\item{t}{Integer. Number of time periods}

\item{p}{Double. Probability of connection between ego and alter.}

\item{undirected}{Logical scalar. Whether the graph is undirected or not.}

\item{weighted}{LoWhether the graph is weighted or not.}

\item{self}{Wheter it includes self-edges.}
}
\value{
A graph represented by an adjacency matrix (if t=1), or an array of
adjacency matrices (if t>1).
}
\description{
Follows the G(N,p) model
}
\note{
The resulting adjacency matrix is dense (hence, be careful with the size)
}
\examples{
\dontrun{
# Setting the seed
set.seed(123)

# Generating an directed graph
rand_graph(undirected=FALSE)

# Comparing P(tie)
x <- rand_graph(1000, p=.1)
sum(x)/length(x)

# Several period random gram
rand_graph(t=5)
}
}

