% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recode}
\alias{recode}
\alias{recode.data.frame}
\alias{recode.matrix}
\title{Recodes an edgelist such that ids go from 1 to n}
\usage{
recode(data, ...)

\method{recode}{data.frame}(data, ...)

\method{recode}{matrix}(data, ...)
}
\arguments{
\item{data}{Edgelist as either a matrix or dataframe with ego and alter}

\item{...}{Further arguments for the method (ignored)}
}
\value{
A recoded edgelist
}
\description{
Recodes an edgelist such that ids go from 1 to n
}
\details{
Required for using most of the package's functions, as ids are used
as a reference for accessing elements in adjacency matrices.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for recoding data frames. It keeps the class of the
object as it returns a data.frame.

\item \code{matrix}: Method for matrices. Likewise the method for data frames,
it returns an object of the same class as the input, a matrix.
}}
\examples{
edgelist <- cbind(c(1,1,3,6),c(4,3,200,1))
edgelist
recode(edgelist)
}
\seealso{
\code{\link{edgelist_to_adjmat}}
}
\keyword{misc}

