% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graph.R
\name{rewire_graph}
\alias{rewire_graph}
\title{Rewires a graph}
\usage{
rewire_graph(graph, p, both.ends = FALSE, self = FALSE, multiple = FALSE,
  undirected = getOption("diffnet.undirected"))
}
\arguments{
\item{graph}{Any class of accepted graph format (see \code{\link{netdiffuseR-graphs}})}

\item{p}{Numeric scalar. Set the probability of changing an edge.}

\item{both.ends}{Logical scalar. When \code{TRUE} rewires both ends.}

\item{self}{Logical scalar. When \code{TRUE}, allows loops (self edges).}

\item{multiple}{Logical scalar. When \code{TRUE} allows multiple edges.}
}
\description{
Rewires a graph
}

