% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graph.R
\name{rgraph_ba}
\alias{rgraph_ba}
\title{Barabasi-Albert model}
\usage{
rgraph_ba(m0 = 1L, m = 1L, t = 10L, graph = NULL)
}
\arguments{
\item{m0}{Integer scalar. Number of initial vertices in the graph.}

\item{m}{Integer scalar. Number of new edges per vertex added.}

\item{t}{Integer scalar. Number of time periods (steps).}

\item{graph}{Any class of accepted graph format (see \code{\link{netdiffuseR-graphs}}).}
}
\value{
A graph.
}
\description{
Barabasi-Albert model
}
\details{
Creates an undirected random graph of size \code{t + m0}.
}
\examples{
# Using another graph as a base graph
graph <- rgraph_ba()
graph

graph <- rgraph_ba(graph=graph)
}
\seealso{
Other simulation functions: \code{\link{rgraph_er}}
}
\concept{
Scale-free random graph

Barabasi-Albert model

Random graph
}
\keyword{distribution}

