% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_equiv.R
\name{struct_equiv}
\alias{print.diffnet_se}
\alias{struct_equiv}
\title{Structural Equivalence}
\usage{
struct_equiv(graph, v = 1, inf.replace = 0, groupvar = NULL,
  mode = "out", ...)

\method{print}{diffnet_se}(x, ...)
}
\arguments{
\item{graph}{Any class of accepted graph format (see \code{\link{netdiffuseR-graphs}}).}

\item{v}{Numeric scalar. Cohesion constant (see details).}

\item{inf.replace}{Numeric scalar scalar. Replacing inf values obtained from \code{\link[igraph:distances]{igraph::distances}}.}

\item{groupvar}{Either a character scalar (if \code{graph} is diffnet), or a vector of size \eqn{n}.}

\item{mode}{Character scalar pased to \code{\link[igraph:distances]{igraph::distances}}}

\item{...}{Further arguments to be passed to \code{\link[igraph:distances]{igraph::distances}} (not valid for the print method).}

\item{x}{A \code{diffnet_se} class object.}
}
\value{
If \code{graph} is a static graph, a list with the following elements:
\item{\code{SE}}{Matrix of size \eqn{n\times n}{n * n} with Structural equivalence}
\item{\code{d}}{Matrix of size \eqn{n\times n}{n * n} Euclidean distances}
\item{\code{gdist}}{Matrix of size \eqn{n\times n}{n * n} Normalized geodesic distance}
In the case of dynamic graph, is a list of size \code{t} in which each element
contains a list as described before. When \code{groupvar} is specified, the
resulting matrices will be of class \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}},
otherwise will be of class \code{\link{matrix}}.
}
\description{
Computes structural equivalence between ego and alter in a network
}
\details{
Structure equivalence is computed as presented in Valente (1995), and Burt (1987),
in particular

\deqn{%
SE_{ij} = \frac{(dmax_i - d_{ji})^v}{\sum_{k\neq i}^n(dmax_i-d_{ki})^v}
}{%
SE(ij) = [dmax(i) - d(ji)]^v/[\sum_k (dmax(i) - d(ki))^v]
}

with the summation over \eqn{k\neq i}{k!=i}, and \eqn{d_{ji}}{d(ji)}, Eucledian distance in terms of geodesics, is defined as

\deqn{%
d_{ji} = \left[(z_{ji} - z_{ij})^2 + \sum_k^n (z_{jk} - z_{ik})^2 +  \sum_k^n (z_{ki} - z_{kj})^2\right]^\frac{1}{2}
}{%
d(ji) = [(z(ji) - z(ij))^2 + \sum_k (z(jk) - z(ik))^2 +  \sum_k (z_(ki) - z_(kj))^2]^(1/2)
}

with \eqn{z_{ij}}{z(ij)} as the geodesic (shortest path) from \eqn{i} to \eqn{j}, and
\eqn{dmax_i}{dmax(i)} equal to largest Euclidean distance between \eqn{i} and any other
vertex in the network. All summations are made over \eqn{k\not\in \{i,j\}}{k!={i,j}}

Here, the value of \eqn{v} is interpreted as cohesion level. The higher its value,
the higher will be the influence that the closests alters will have over ego (see
Burt's paper in the reference).

Structural equivalence can be computed either for the entire graph or by groups
of vertices. When, for example, the user knows before hand that the vertices
are distributed accross separated communities, he can make this explicit to
the function and provide a \code{groupvar} variable that accounts for this.
Hence, when \code{groupvar} is not \code{NULL} the algorithm will compute
structural equivalence within communities as marked by \code{groupvar}.
}
\examples{
# Computing structural equivalence for the fakedata -------------------------
data(fakesurvey)

# Coercing it into a diffnet object
fakediffnet <- survey_to_diffnet(
   fakesurvey, "id", c("net1", "net2", "net3"), "toa", "group"
)

# Computing structural equivalence without specifying group
se_all <- struct_equiv(fakediffnet)

# Notice that pairs of individuals from different communities have
# non-zero values
se_all
se_all[[1]]$SE

# ... Now specifying a groupvar
se_group <- struct_equiv(fakediffnet, groupvar="group")

# Notice that pairs of individuals from different communities have
# only zero values.
se_group
se_group[[1]]$SE



}
\author{
George G. Vega Yon, Stephanie R. Dyal, Timothy B, Hayes, Thomas W. Valente
}
\references{
Burt, R. S. (1987). "Social Contagion and Innovation: Cohesion versus
Structural Equivalence". American Journal of Sociology, 92(6), 1287–1335.
\url{http://doi.org/10.1086/228667}

Valente, T. W. (1995). "Network models of the diffusion of innovations" (2nd ed.).
Cresskill N.J.: Hampton Press.
}
\seealso{
Other statistics: \code{\link{classify_adopters}},
  \code{\link{cumulative_adopt_count}}, \code{\link{dgr}},
  \code{\link{exposure}}, \code{\link{hazard_rate}},
  \code{\link{infection}}, \code{\link{moran}},
  \code{\link{threshold}}
}
\keyword{univar}

