% Generated by roxygen2 (4.0.2): do not edit by hand
\name{threshold}
\alias{threshold}
\title{calculates threshold}
\usage{
threshold(exposureObject, toa, net_no, conv)
}
\arguments{
\item{exposureObject}{exposure matrices from ExposureCalc}

\item{toa}{}

\item{net_no}{number assigned to network (might remove)}

\item{conv}{convention number (might remove)}
}
\description{
calculates threshold
}

