% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{threshold}
\alias{threshold}
\title{Calculates threshold}
\usage{
threshold(exposure, times, times.recode = TRUE)
}
\arguments{
\item{exposure}{\eqn{n\times T}{n * T} matrix. Exposure to the innovation obtained from
\code{\link{exposure}}.}

\item{times}{Integer vector. Indicating the time of adoption of the innovation.}

\item{times.recode}{Logical. TRUE when time recoding must be done.}
}
\value{
A vector of size \eqn{n} indicating the threshold for each node.
}
\description{
Threshold as the exposure of vertex by the time of the adoption (see \code{\link{exposure}}).
}

