% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{toa_diff}
\alias{toa_diff}
\alias{toa_diff.integer}
\alias{toa_diff.numeric}
\title{Difference in Time of Adoption (TOA) between individuals}
\usage{
toa_diff(times, recode = TRUE, ...)

\method{toa_diff}{integer}(times, recode = TRUE, ...)

\method{toa_diff}{numeric}(times, recode = TRUE, ...)
}
\arguments{
\item{times}{Integer vector containing time of adoption of the innovation.}

\item{recode}{Logical. When TRUE recodes time (see details).}

\item{...}{Ignored.}
}
\value{
An \eqn{n \times n}{n * n} symmetric matrix indicating the difference in times of
adoption between each pair of nodes.
}
\description{
Creates \eqn{n \times n}{n * n} matrix indicating the difference in times of adoption between
each pair of nodes
}
\details{
Each cell ij of the resulting matrix is calculated as \eqn{toa_j - toa_i}{%
toa(j) - toa(i)}, so that whenever its positive it means that the j-th individual (alter)
adopted the innovation sonner.
}
\examples{
# Generating a random vector of time
set.seed(123)
times <- sample(2000:2005, 10, TRUE)

# Computing the TOA differences
toa_diff(times)
}

