% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{toa_mat}
\alias{toa_mat}
\alias{toa_mat.numeric}
\title{Difference in Time of Adoption (TOA) between nodes}
\usage{
toa_mat(x, ...)

\method{toa_mat}{numeric}(x, ...)
}
\arguments{
\item{...}{Ignored}

\item{time}{Integer vector with times of adoption}
}
\value{
An n x n matrix indicating the difference in times of adoption between
each pair of nodes.
}
\description{
Creates n x n matrix indicating the difference in times of adoption between
each pair of nodes
}
\examples{
# Generating a random vector of time
set.seed(123)
times <- sample(2000:2005, 10, TRUE)

# Computing the TOA differences
toa_mat(times)
}

