% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.R
\name{toa_mat}
\alias{toa_mat}
\alias{toa_mat.integer}
\alias{toa_mat.numeric}
\title{Time of adoption matrix}
\usage{
toa_mat(times, recode = TRUE, ...)

\method{toa_mat}{numeric}(times, recode = TRUE, ...)

\method{toa_mat}{integer}(times, recode = TRUE, ...)
}
\arguments{
\item{times}{Integer vector containing time of adoption of the innovation.}

\item{recode}{Logical. When TRUE recodes time (see details).}

\item{...}{Ignored.}
}
\value{
A list of two \eqn{n \times T}{n x T}
 \item{\code{cumadopt}}{has 1's for all years in which a node indicates having the innovation.}
 \item{\code{adopt}}{has 1's only for the year of adoption and 0 for the rest.}
}
\description{
Creates two matrices indicating the time of adoption (TOA) of the innovation, and times during
which an individual had already adopted the innovation.
}
\details{
By construction this function requires time units to be between 1 and T, where
T is the length (number) of time periods. By default a recoding of time is
performed as \eqn{time'=time - min(time) + 1}{time'=time - min(time) + 1}.
}
\keyword{manip}

