% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{transmission_tree}
\alias{transmission_tree}
\title{Retrieve the transmission tree of a \code{diffnet} object}
\usage{
transmission_tree(x)
}
\arguments{
\item{x}{A \code{diffnet} object.}
}
\value{
A \code{data.frame} with columns \code{date}, \code{source},
  \code{target}, \code{source_exposure_date}, \code{virus_id}, \code{virus}.
}
\description{
Returns the data.frame stored in \code{x$transmission$tree}. If none has
been attached, an empty data.frame with the standard columns is returned.
}
\seealso{
\code{\link{as_transmission_tree}}
}
